Config: Deprecate most fields, and hook into ConfigBuildFlags
Copy 3.xml to 4.xml, and patch 4.xml to deal with the fact that
Config.DEBUG's value is no longer actually fixed.  We don't want
to modify 3.xml directly, because it's the official cupcake API.

Update current.xml to reflect the differences between 3.xml and 4.xml
and to reflect the new deprecation states.

The deprecated fields should be @hidden if possible before the
next API release, but it can't be done yet because there are
SDK-linked platform apps that use some of the deprecated fields.

Signed-off-by: Dave Bort <dbort@android.com>