Move ContactsContract.java and SocialContract.java into android.providers
The ContactsContract and SocialContract lived previously in the
com.android.providers.contacts2 package, and could not be accessed by
other packages from there without symlinks, which was getting messy. If
it turns out there was a good reason for having the contracts in that
package we may have to move these back, but for now this seems like the
obvious configuration.