Remove obsolete notion of invalid node.
In the old navigation model, we set focus to a node, and checked to see
if that node was still valid (could have changed since we built the
navigation cache).  Then we sent a message back to the UI thread to
tell it to draw the ring differently.  In the new model, we are essentially
moving a mouse, so we do not concern ourselves with whether the node is
valid.  Removed all the code for passing the message to the UI thread
and calling into native.  Needs a change to external/webkit to work
properly.