NEW_API: Expose queryAfterZeroResults searchable attribute.
If this attribute is provided and true, the searchable activity will be invoked for all queries in a particular session.
If set to false and the activity returned zero results for a query, it will not be invoked again in that session for supersets of that zero-results query. For example, if the activity returned zero results for "bo", it would not be queried again for "bob" as an optimization.

The default value is false.