Fix intermittent crash related to call waiting
Previously, when call waiting is ignored by user or timed out
by PhoneApp, the connection object for call waiting does not
release the wake lock. It causes the phone process to crash
randomly.

The fix is to release the wake lock when call waiting is
ignored by user or timed out by PhoneApp.