Allow proper text editing after clicking with the trackball.
As with the recent change to make a touch bring up the WebTextView,
so that the trackball can be used, and we get text editing goodness,
we also need to bring up the WebTextView when clicking the trackball.
Also do not call overrideUrlLoading for an input field and add a
comment.