Fixes for moving a focused textfield on screen when entering a character.
When the focused node is a textfield, but the WebView has focus
(for example if the page had an initial focus) and the cursor
is not over a plugin or textfield, rebuild the WebTextView
to handle the keys.  Also, remove mScrollToAccommodateCursor,
since now the WebTextView is only showing when we want to
scroll.