Make sure SearchDialog.mSearchable != null in key handlers
This change adds checking that the searchdialog is initialized
properly to a few key event listeners where it was missing
before.

Also makes sure that the search dialog is not shown if the
global search is requested but the global search provider
cannot be found.

Should fix http://b/issue?id=1982128
"Device rebooting after search results"
where SearchDialog.onKeyDown() threw an NPE