Re-add double tap to the Browser. It only works if WebView uses wide viewport. It is also controlled by ENABLE_DOUBLETAP_ZOOM flag. Currently the default is on.
Double tap will toggle between current zoom level and all the way zoom out,
in another word, zoom overview mode.

While in zoom overview mode, don't display zoom control while panning. You
can click to follow the link. If clicking on a text input box, it should
bring it back to the normal mode so that it will be easy to edit text.

While in zoom overview mode, the page is zoomed to fit the width in the
view. While the text paragraph are wrapped to the matching normal mode scale.
So there is no reflow when we switching between two modes.