Fix 2016808. We used to send the new scale/scroll info to the UI thread after WebCore finish the first layout. As the picture in the UI thread is still the previous one, if there is a UI update happened before we generate the new picture, the new scale/scroll will be applied to the old picture.
On the other hand, we do want to trigger the resize for the new page after
the first layout. e.g. The page was at nytimes.com. Now user types google.com.
google.com is first layout in the wide viewport where nytimes.com is. We dont
want to show that picture. We would like to show the fit to screen version of
google.com as the first picture.

Remove the SnapToAnchor code which is left over from previous zoom attempt.