Pass clicks to webkit's textfield to determine the new selection.
Requires a corresponding change to external/webkit.  We were previously
letting WebTextView handle clicks, determine the change in selection,
and pass that down to webkit.  This sometimes resulted in a different
placement of the caret if the WebTextView and the webkit-rendered
textfield did not line up exactly.  Now, we pass the click directly
to webkit, which determines the new selection and passes that info
back to the WebTextView.  This also has the benefit of letting
the WebTextView reflect changes in the selection that originated from
webkit. Also remove some unused parameters.