Add rotary based lock screen, and only show security pattern screen when pattern is set.
When a pattern is enabled, there is no informational lock screen, just the pattern
screen.  Eventually we will get a new layout that manages to show the relevent information
(date / time / charging info / alarm) on the pattern screen.

Also: I noticed that when there is no backup option available, we still showed
the "forgot pattern?" button, which would do nothing.  I fixed that; we
no longer show the "forgot pattern" button if there is no account unlock
option available.

a