Deprecate showSearchIconAsBadge and icon search attrs
mlebeau says:
"Basically, because we now show the app icon for in-app
search, showSearchIconAsBadge is a bit unnecessary, but it does in
fact still work. Basically if either showSearchIconAsBadge or
showSearchLabelAsBadge are specified, we grow the height of the search
bar and place the provided icon or the provided label, respectively,
above the text field and app icon. But we do not remove the app icon
if the developer specifies showSearchIconAsBadge - it's still just the
'badge area' above the icon."

Fixes:
http://b/issue?id=2105329
"Deprecate searchable attributes 'showSearchIconAsBadge' and 'icon'"

Change-Id: Ie2976aafe42b4ab870be9e64b34eb0ed441ebbb0