Fix call collision issue
This patch is to fix the call collision issue between an MT call
and an MO call causing the MT call screen shown with the MO call info.
The fix is to add the condition of foreground call to determine the
call collision on top of the pendingMO call only.

Bug ID: 2116865
Change-Id: I73aa310d39ebea97e6183281821bf941b5890ccf