Cherry-pick from mr2 to eclair: DO NOT MERGE
commit 149c0543a3d5f450686b7c3d142ac5f8fcf61ed4

BUG=2176949

Partial fix for 2069703: Fix race condition in KeyguardViewManager.hide() that could result in a blank screen.

This fixes a race condition seen whenever show() is called before the timer in hide() triggers. This can happen
for example if the user hits "Emergency Dial" followed by the back button before the 0.5s timeout completes. The result
is a blank screen with no keyguard view and no means to recover on devices w/o a keyboard.

The bug caused us to sometimes remove the newly created KeyguardView instead of the old one, leaving the view empty.

The fix is to always remove the last view.