Cherry-pick into donut:   DO NOT MERGE
commit a9f1dd021f8f6ee777bc4d27913bd40c42e753af

Make <font size> and <font height> in string resources respect density.

TextPaint gains a "density" field so that TextView can pass the density
along.  AbsoluteSizeSpan gains a new flag to indicate that its argument
is in dip instead of in physical pixels.  LineHeightSpan gains an inner
interface whose chooseHeight() method includes a TextPaint argument so
it can get at the density.  And when StringBlock creates the markup
objects, it now uses the density-aware versions.

These are all @hidden here since they are not part of the donut API.

Bug 2174236

Change-Id: I61b20d961707768d14387c37dd41ae38b21e7c7d