telephony/cdma: Fix Erroneous Roaming Indicators and Latitude-Longitude parsing
When Android Telephony receives response to GET_REGISTRATION_STATE
message from RIL it may contain few fields set to NULL. Due to a parsing
exception encountered while parsing that field, the remaining fields will
not be parsed even if they are valid data. Ignore all fields that are NULL
while letting it parse non-NULL data.

For Latitude and Longitude, the values should not be hexadecimal. They
shall be parsed as decimal values as specified in the 3GPP2 C.S0005-A
specification. Invalid value is changed from -1 to Integer.MAX_VALUE.

Bug: 2201613
Change-Id: I13dd02fcfa2ae7fcb6f21c4b94b830786bd7270c