Send "encoding problem" to the network for general errors during SMS dispatch.
Previously we were returning "other terminal problem" to the network, but this
had the potential of blocking MT SMS messages for days at a time (eg, if the
problematic message is resent).  Sending the "encoding problem" cause code is
reasonable since in most cases the we have encountered an error while trying
to parse the message.

Addresses b/2200412.