Fix #2262593: 42 ANR reports from android.process.acore
There was a lock in AssetManager that purported to be per-instance (mSync)
but was actually static.  A lot of code used it like it was per-instance,
but this would actually block all instances.  This is now changed to fix
the name and make everything except the actual static data lock on the
specific AssetManager instance.

Change-Id: Ie8e9ad60f962184e76b2301f7a2790d0c2487063