Fix soft keyboard dismissing itself when zooming into a textfield in browser.
We remove the WebTextView while zooming.  If the WebTextView ended up
offscreen, we remove it completely, thus hiding the IME.  In some cases,
however, the WebTextView is only offscreen because the IME overlapped it.
Check to see if the IME is showing, and if so, adjust the WebTextView and
let it get scrolled into view.

Also perform the removal/change in text size inside the method, so it
happens in both places.

Lastly, do not call didUpdateTextViewBounds if there is no WebTextView.

Fix for http://b/issue?id=2266066