Only hold a weak pointer on SurfaceComposerClients
Each process maintains an array of active SurfaceComposerClient
objects, so that they can be reused as new surfaces are parceled
across.  When a SurfaceComposerClient is disposed, it will remove
itself from this list.  However, because the list maintains a strong
reference on the object, a reference cycle is created, and the
client is never deleted.

This patch changes the list to maintain weak pointers on the clients
instead.

Change-Id: I93dc8155fe28b4e350366a3400cdf22a8c77cdd3