Fix 2468960: Make sure unlock screen gets updated when SIM state changes.
Added code to let us lazily re-create the unlock screen when the SIM
status changes.  This fixes a bug where we sometimes show the Pattern/Password unlock
instead of the SIM PIN unlock due to an ordering issue with recent changes
to the telephony layer SIM update logic. It now correctly re-evaluates the SIM state
and updates the UI accordingly.