Fix for deadlock between StatusBarService and NotificationManagerService
A ServerThread holding a lock on mQueue in StatusBarService invoked a
callback in NotificationManagerService which required a lock on
mNotificationList. At  the same time, a BinderThread holding a lock on
mNotificationList was attempting to post a message to StatusBarService
which requires lock on mQueue. The fix is to release the lock on mQueue
in handleMessage() before running the actions at the end of the method.