Corrected debug message in WindowManager
When WindowManager reports "Key dispatching timed out"
it prints out information about the window state that was
present at the time the key was sent to that window.
There is a minor error in the class representing the
recorded window state so that the currently focused window
is printed instead of the recorded focused window.

Change-Id: I29a5471ef725e30f812ffd57fd4597ce81c0c7f2