Download manager support for file URIs + last modified time
First, in the download manager public API, I'm changing
COLUMN_REQUESTED_TIMESTAMP to COLUMN_LAST_MODIFIED_TIMESTAMP, a field
that already exists in the download manager.  Upon inspection of
existing users of this field, it looks like there are some good use
cases -- for instance, looking at when a download completed.  And if
I'm exposing that timestamp, there's no need to also expose the
requested timestamp.

Second, this change includes support for file URI destinations:

* add DESTINATION_FILE_URI for use by the DownloadManager wrapper
* make DownloadManager use it appropriately

The backend support is in a separate change (since it's in a separate
project).

Also fixing a bug with file URI construction when DownloadManager
fills in COLUMN_LOCAL_URI.

Change-Id: I5bb93a222cd23215c2a03ed9f68d7f140454f7f2