Fix a merge problem with html mms
Html mms message support was added back in Jan '10. At that time, we
had moved the mms code out of the framework into the mms app. We decided
to back out that change and leave the mms code in its original place.
As a result, the changes to support html messages were lost. This
handmerged CL restores those changes. I'll cherry-pick this into master
as well. Bug 2858888

Change-Id: Icf8835edc8ac396698c167be5433a6fe1cfe6103