DO NOT MERGE -- copy on tapping outside of selected text in gmail
Text can be selected in gmail by choosing Select Text from the
menu and dragging the selection. There's no pretend action bar
like there is in the Browser to commit the selection.

When tapping outside of the selection, copy the text, if
the client (like gmail) has mapped the trackball to the
arrow keys.

Change-Id: I7dd133fd860d392bb01ec103e379e1f1d2a43e64
http://b/2626451