DO NOT MERGE Tweaked EdgeGlow to better match user expectations.
Change-Id: I7864dd4d3e15a9dc11429d5d3268f9b078f3eaad
Constants updated to match user testing.
Recede no longer flashes out but degrades all the way to zero.
Growth of the absorb flash is now quadratic as this better
matches the "physics" of the user flicking the list.
Size of the pull and absorb graphics changed to give the user
a more dramatic indication of what has happened.