Implement contacts lookup (caller-id) for incoming SIP calls.
CallerInfoAsyncQuery can now handle SIP addresses in addition to regular
phone numbers: if the number passed in to startQuery() is actually a "URI
number", we now treat it as a SIP address and look it up directly in the
Data table.

If it's a regular phone number, the behavior is unchanged: we use the
PhoneLookup table as before.

This piece of the fix covers only the contact lookup for incoming calls;
we still need some more cleanup of the CallerInfo class in order to get
the call log working.

Bug: 3004127

Change-Id: I0fcb80f9de5b8ecf99d31ee92e0889ddb07216fd