Start of input method control on system bar.
This change creates a new icon to the right of the
notification area for the IMM to expose control over the
current IME. Currently it shows the IME subtype picker.

TODO: Connect directly to the IMMService to show the correct
icon (and hide the icon entirely when the input panel is not
visible).

Bug: 2975425
Change-Id: If07f30867b81950a0a86b00807a41e2ff2af389c