Fix bug 2964954 - Lists blow up if their adapter changes while detached.
If a list adapter's content changes while the ListView is detached
from a window, the data change is not registered and the next
layoutChildren call causes the ListView to detect a bad change that
the adapter did not notify ListView about. Any time a ListView becomes
attached to a window, has a valid adapter, and no attached data set
observer, treat it as a data set changed event.

Change-Id: Ic00c05a1a0d063399dc78d5d877e3d5f9b1bee16