Work around race condition in CallBackProxy.
If WebCore requests two (or more) alert/prompt/confirm dialogs
in quick succeession, we have a race that when the dialog is dismissed
the WebView does not have it's focus restored before the next dialog
is requested to be shown and we early out of displaying the dialog (i.e.
we do not show it).

Work around this by no longer verifying that the Activity has focus as
this check seems to be redundant anyway (see 3166409).

Bug: 3151825, 3166409

Change-Id: I7f650b066ee63da92bc9a620b8e190c91b306d3e