StrictMode.Span annotations for ScrollView.
Between the animation system, ListView, and this, we can now account
for the large majority of animations in-flight at any point and annotate
StrictMode violations accordingly.

Note that StrictMode.Span goes out of its way to not allocate memory
and is very light-weight, especially on user builds where it's a
no-op.

I've tested this with the LOG_V debugging in StrictMode and I think
I've gotten all the entry/exit cases of ScrollView's animations.  At
least, I'm not able to get it out-of-sync at least, even with
orientation changes during scroll/fling, etc.  It's possible I missed
a path, but that should show up as excessive false positives during
dropbox uploads if so.

Change-Id: I3b0545111a775d9fa8e171f2a190031f1b56c4c0