Get rid of setupDataCallWithProtocol.
setupDataCallWithProtocol was introduced to introduce a
protocol parameter to setupDataCall without having to
modify its callers. This is not very useful, as there are
only two callers. Get rid of it and make setupDataCall take
the protocol parameter itself.

For now, hardcode IPv4 connectivity. When we add the
protocol field to ApnSettings, it will be fetched from
there.

Change-Id: I1a1adc616ffd9ac68be6911f054790da48e472d6