Avoid superfluous calls to CacheManager with the Chromium HTTP stack
When using the Chromium HTTP stack, most of the calls to CacheManager
methods are not required, as we're not using its cache. These methods
are now marked with asserts to make this clear.

Also avoid creating the cache database in WebViewDatabase.

This will avoid creating the database databases/webviewCache.db and
the directory cache/webviewCache.

Bug: 3270236
Change-Id: I68f94dde16830ed817454d5e1af961f41b71d018