Fix spurious resizes during rotation animation.
Fixed a bug in WindowManagerService where it would set mSurfaceResized
to true even if the surface was just moved and not resized.  As a result,
we would send dozens of spurious resize messages to all applications
during orientation changes since the rotation animation changes the
positions of surfaces as they swing into place.  Among other things,
the spurious resizes caused the wallpaper to be redrawn dozens of time.

Change-Id: I2be7278c0d6a00aaef665a65e268d8da1771e51f