Add byte limit in addition to ratio for installs DO NOT MERGE
Previously you were not allowed to install anything to the internal
storage if there was less than 10% free space available on the
internal storage. This doesn't make sense on devices with huge storage
capacity, so just add in a hard-coded byte limit.

This will be cleaned up at a later time to use the same code as
DeviceStorageMonitorService to determine the limits.

Bug: 4086510
Change-Id: I39f69b5272ef44fea218e3b45171b48e2eab2480