OBEX: Fix PrivateOutputStream small write problem
When data less than max packet size in length is sent into the write
method the data will only be added to the internal buffer. If several
calls to write is performed by the application continueOperation will
not be called at all. The solution to the problem is to always check
the internal buffer size and to call continueOperation every time
maxPacketSize bytes is in the internal buffer.

Change-Id: I5ebfa3c26db2c1aefe1a115d7782d8ceaa760937