Nicer API for disabling/enabling the home/up button in action bars.
Have an explicit call for enabling the home/up button. Auto-enable it
if the app targets < ICS to preserve Honeycomb behavior. Auto-enable
it if the app shows home as up.

This prevents unwanted touch/focus feedback on the home button when
the app hasn't wired it up to do anything useful.

Change-Id: Icfe95ab5a11b3998bca08b0fbbfe1bf6c3c89b5d