Bug 4382356: No automatic focus change in ScrollViews
The current behavior in ScrollView is inconsistent: a slow scroll will
not change focus while a fling will. With this change, the focus is not
moved from field to field (resulting in poup due to keyboard / suggestion bar
state changes).

Also made sure any new input in a TextFields scrolls back the view to make the
EditText visible. Has no effect if the EditText is already visible.

Change-Id: I7ff301fe85f73c27d76e4893aeab7442d0539b13