Removed READ_WRITE_ALL_VOICEMAILS permission from framework.
This permission is meant to be a 'signature' permission and to be used
only by the contacts app. We recently moved it to framework/base as part
of opening up voicemail api into the SDK. However, the signature of
framework is different from contacts app. Consequently the contacts app
is not granted READ_WRITE_ALL_VOICEMAILS permission. This makes the
contacts app crash on start.

This fix removes the READ_WRITE_ALL_iVOICEMAILS from framework/base so
that the one defined in contacts provider is then used by the system.

Bug: 5054221

Change-Id: Iaa7a23fe95b0142978eff124e25fe2291dc7d212