libstagefright: add destroyOMXPlugin and use C linkage for symbols
Add a destroyOMXPlugin() to complemenet createOMXPlugin().  Since the latter is
an opaque call into a library, it is not safe to assume that the object thus
returned may be destroyed by calling delete.  This patch will call
destroyOMXPlugin() in libstagefrighthw.so, if the symbol is defined there.
Otherwise, it will default to a delete as before.

Also, prefer C linkage for the symbol names.  Using mangled C++ symbols is not
very safe and makes it hard to grep for those symbols through the code.  We
attempt to locate createOMXPlugin as a C symbol then try it as a mangled C++
one.  We only use C linkage for destroyOMXPlugin since this symbol is new.

Change-Id: Ied23c910486856a0f18056df049c871234797c5c
Signed-off-by: Iliyan Malchev <malchev@google.com>