Warn only if we're tearing down "live" DeathRecipient content
If the native-side bookkeeping still has strong references to VM-side
DeathRecipient objects at the time when it's being torn down, that
suggests that the app is doing unwholesome.  Log a warning to that
effect, with the class name of the objects to try to help the developer
figure out what they're mishandling.

Fixes bug 5202777 -- in particular, it no longer logs in the
working-as-intended case following delivery of the death notices,
when we've got the existing list shell but the weak refs have properly
cleared.  Also step down from "error" to "warning" logging as befits
the nature of the actual situation now being described.

Change-Id: Ic393560824800fbd912a6e69692c65be4fcc7544