Add "format" extra to SMS delivery report intents.
The MMS app was updated for multiple SMS dispatcher support, passing
the "format" extra into SmsMessage.createFromPdu() so that the PDU
is interpreted as the correct format (3GPP or 3GPP2).

This change adds the "format" extra to the intents that are sent for
SMS delivery reports. The MMS app is already expecting this extra, so
this change fixes a NPE that was introduced with the last change.

Bug: 5434052
Change-Id: I38b97f166c78d246d721ca1db8a4a18a2c6529bd