Fix 5369428: Use full battery state information when determining charge status
This is a fix for a bug where we'd show "connect charger" when the device was
connected but not charging due to the battery being in an unchargeable state
(too hot, cold, etc).

It now maintains a full copy of the battery state and uses the plugged status
to determine if the device is plugged in.

Change-Id: I60fa4e4566a45663b130f0ff4863bcc595ae3c4a