NumberPicker incorectly enforcing minimal width and height.
The min width and height were enforced with an exception which
is not correct since a view not having enough space should not
terminate the application. Now the insufficient vertical or
horizontal space is propagated by setting the
MEASURED_STATE_TOO_SMALL flag and respecting the constraints
imposed while measuring (consistent with the rest of the View
hierarchy).

bug:5572237

Change-Id: I2dbeb5451d3426cc983b41f6023c7b0fc52c7da3