Added Face Unlock to tablet lockscreen layouts
The Face Unlock Area was not part of the tablet layouts, so prior to
this change, Face Unlock would not show up on tablets when selected as
the unlock method.  The backup unlock method would show up instead.

The goal here is for the pattern and PIN unlock layouts (in both
portrait and landscape mode) to look the same as before this change.
This was a little harder than it was with the phone layouts for two
reasons:

1) For the phones it was ok for Face Unlock to be sized such that it
just covers the backup method.  For the tablets we want Face Unlock
to cover far more real estate.

2) The phones were based on a grid layout, whereas the tablet is a
linear layout.

Note that the diff makes the modifications look way more extensive
than they actually are.  Basically, in most cases I am putting a
relative layout around some existing portion of the layout and
putting the Face Unlock Area area inside of the new relative layout.

Change-Id: I478becddf2a9ee9fe7b6d653e604fa3ad89b822f