CrespoMMS appears after delay in the thread
Bug 5032682

Part A of several parts to speed up sending MMS's. In this change, don't
allow multiple threads to write out the same pdu at the same time. This
was causing problems between the thread sending the mms message and the
UI trying to refresh and display the in-progress-sending message.

Change-Id: I862081619e6b6808caaba86a3b48820e0ef75aba