Improve AudioPlayer position reporting
The latency was not taken into account when updating mPositionTimeRealUs
inside of the fillBuffer hook, contrary to what the getRealTimeUsLocked()
method does. This caused the realTimeOffset calculated in the
getMediaTimeUs to always be negative, causing the reported position to
always be equal to mPositionTimeMediaUs, which is updated infrequently.
With this change, the reported position is updated more frequently, allowing
apps to perform smoother UI updates.

Change-Id: I61e05c1a8b53d46b9091afb0d18a6289d13a7a5e