Set AuthType to NONE if username of APN is not set
When APN is edited, if authType and username are both not
set, authType should be set to NONE before setup of
data call.

The code today checks mApn.user for NULL, but if
user has edited APN the mApn.user field will be set
to an empty string.

Change the check of mApn.user to also take care of
this case.

Change-Id: I0e49247d46e1ba93da6dd951bbb75acc63fac73f