The beginning of expanded notifications.
There are now two "rebuilder" classes, each of which
consumes a Notification.Builder and modifies its behavior.
(Inheritance in Builder classes is...not advisable.)

- BigPictureStyle: includes a large Bitmap above the usual
  notification strip.

- BigTextStyle: shows the contentText in a large, wrapping
  TextView instead of truncating to one line.

As for SystemUI, the notification panel now shows the
expanded form if it is available, otherwise the usual
contentView is shown.

(Note that the structure of largeIcon notifications has
changed a bit: The largeIcon is no longer handled by the
status bar at all; it's entirely inside the template now.
Not only does this make the code simpler, and make large
notifications possible, but it fixes the longstanding
irritation that tapping on a largeIcon doesn't highlight the
whole notification row. Man, that feels good.)

Change-Id: I2b9d8a6ea4385659d8cb1ed467c1caf5e12628dd