Remove deprecation on android:enabled attribute.
The android:enabled was never deprecated and replaced by
android:state_enabled as it was mentioned in the documentation.

The attribute android:enabled when applied to a TextView (or its
descendants) still works until now. In fact that's the way to
disable widgets from layout XMLs.

The deprecation made ADT and lint suggest users not to use this
attribute.

Change-Id: I5e7d7060d9e79a04342ebc723d7937b9bc12a018
Signed-off-by: Yuku on yuku4 <yukuku@gmail.com>