Fix: refreshBatteryInfo was uncovering backup
The onRefreshBatteryInfo() function hides the biometric sensor area so
the biometric sensor isn't shown when the screen turns on during a
plug/unplug.  However, sometimes the function is called when there
is no plug/unplug.  This fix makes it only hide the biometric sensor
area (thus exposing the backup lock) on plug/unplug.

Note that this does not fix b/5465535 although it may help reduce its
frequency.

Change-Id: Iccff382c23d71257593dc1595ed8fec57adaf5bd