Fix 6396479: Made FUL unlock layout look good on nakasi
For the portrait layouts, the size of the unlock view was reduced
because it was outside the bottom bounds of the screen, making it
look much wider than it was high.  The fix makes it look square.

For the landscape layouts, the size of the unlock view was increased
because it was leaving only thin slivers of background space on the
top and bottom.  It now fills the entire right half of the screen.

This was also tested on the Xoom to make sure it looks reasonable on
10 inch tablets, even though FUL doesn't ship on the Xoom.

Change-Id: I969912a944d20560d31729806273dec47aaa6ecb