DO NOT MERGE
Change BootAnimation's initial display size query to check
a system property instead of the SurfaceFlinger's display
size (which comes from the frame buffer) in case the hwc
choose a display size that is smaller than the full frame
buffer.  The hwc would do this if the monitor's preferred/native
resolution is less than 1080p.

Change-Id: I33bf06fe06a21e778cf3f23eaf0374346a6e2aa0
Signed-off-by: Mike J. Chen <mjchen@google.com>