Add --error-on-failed-insert option to aapt. do not merge.
The new SDK build system give the ability to insert
versionCode/Name and min/targetSdkVersion in the manifest
but aapt won't replace those if they already exist.

The main problem is that aapt doesn't actually fail when
it doesn't replace them, making the output not what the
developer wanted.

This patch set adds an option to aapt to make it return
an error if the insert failed because the attribute
already existed.

Cherry-pick from 7714a2429b192c88e134ff67b969121bbaeb5457

Change-Id: I8938ec1238da407a8562c974e9598db39001ffd9