Disable HDMI rotation lock.
We used to lock the rotation to landscape when HDMI
was plugged in because we did not have a mechanism
for independently rotating secondary displays.
Now as part of the multi-display work we do have
a way to do this so we can turn off the special HDMI
rotation lock.

Bug: 7410631
Change-Id: I7948e92ebe6312a1e2eeeb91250259cd67a00fe0