FUL being unsuppressed when screen turns off fixes b/7453702
We suppress FUL when the device first boots because with all of the
things happening at boot, we don't want to interact with the camera.
We also suppress it immediately after a user switch for the same
reasons.  However, once the user turns the screen off, we want to
show FUL when they turn it back on.  Prior to this fix it was only
being unsuppressed on login.

Note that this CL also removes the isFirstBoot flag.  It was only
being set/checked when the suppress flag was set/checked, so it was
redundant.

This also addresses bug b/7450665, which is regarding the
user-switch case.

Change-Id: I2b39e85e21c08ef7629398521ab1159f5a6c7547