Revert partial RTL'ization of Editor.setError()
The conversion of the error indication on Editor to RTL-aware was only
partially completed, and was causing bugs such as an error indication
failing to appear when set (bug 7457897).

This patch reverts these changes and just always sets the error drawable
on the right. This fixes the above bug, and also makes the error
drawable position always consistent with the error popup (before, in an
RTL layout direction, the popup would be on the right and the drawable
on the left).

Making the error display fully RTL-aware should be done as future work.

Change-Id: Icaee91210454ed9056e7200520d9275303de02ca