Use stable layouts for keyguard/SlidingChallengeLayout
When coming back from the secure camera we will be transitioning from
fullscreen back to not-fullscreen with regard to the status bar. Flag
SlidingChallengeLayout to force the window to have stable insets so
that our layout doesn't jump around during this transition.

Set the same flag in MultiPaneChallengeLayout as well so that similar
cases with other activities are also covered.

Bug 7498950

Change-Id: I1b65e04d1404e764ddad29472b70cf791d6801f9