Fix multi-user support in keyguard.
This fixes a bug where user0's credentials were required of all users.

LockPatternUtils used to check for the system process of the caller
to determine which user id to use.  Now that keyguard is in its own
process, this can no longer be used.

The fix is to use the permission system to see if the calling process
is allowed to interact across users and if so, use the id of the current
user.

Change-Id: I2d7111938bb3bf381f72698c9fa4fb848d237153