Fix widget update issue
This fixes an issue where widgets weren't being updated.  It
was caused by creating widgets with the wrong package id.

Keyguard used to be identified with package "android" when it
was running in the system proces.  Now that it's separate, we need
to explicitly pass in the package name for keyguard ("com.android.keyguard")
when we allocate app widget ids.

Change-Id: I48e1b78d97aa3d0271773219e477d3b741994901