Ensure wallpaper creation succeeds
ImageWallpaper was sometimes querying GL for a max texture size too early
(before the first call to eglMakeCurrent()). This problem caused the wallpaper
to then get created in software, resulting in noticeably slower performance
when the wallpaper was visible.

This fix ensures that the
makeCurrent happens before the query, ensuring that wallpapers of the right
size actually get created instead of failing due to this error at creation time.

Issue #8319960 sluggish yakju animations over launcher

Change-Id: I12a3eba9f1818bdf544691e0727fe12f7e820651