Dynamically togglable accessibility features improperly handled.
Touch exploration and enhanced web accessibility can be toggled at
run time. However, the code that updates the state of these features
was not properly doing that. In particular, it did not write to the
settings if the feature gets disabled. Now the logic is much cleaner.
If there is a service that can request and requests a feature and
the feature is not enabled, we enabled it, otherwise the feature is
disabled.

bug:8790771

Change-Id: I218dfa12fd02220c94940b54f42bed578811a794