Move Keyguard to the systemui process.
This should save some memory on devices with limited amounts of RAM
by allowing it to share the graphics context with systemui.

Initially, adding this to the systemui process triggered a bug where
the shared process id didn't exist in the table yet. The change to
ActiveServices tries to work around this by explicitly adding the package in
attachApplicationLocked().

Change-Id: I9e97d5d69ddd15d7f16aca83a57bdf381811a13b