Fix ListMenuItemView regression from patch 8933efd49
Passing 'true' as the third parameter of LayoutInflater#inflate does
cause it to add the inflated view to the parent in a single step, but
it also changes the return value to be the root of the resulting
hierarchy after the add rather than the root of the inflated layout.

Return to adding the view to its parent as a separate step.

Bug 9754101

Change-Id: I7ca4c921993594931f5007af69d45e9ff06065df