Remove unused views when we return to keyguard
This fixes a bug where launching the camera would move
the camera view to the root hierarchy.  In insecure mode, a
subsequent dismiss() comes in and tears down the view hierachy
before the event can be processed, which left the view in the
root hierarchy.

The fix is to remove all children when we reconstruct keyguard.

Fixes bug 10640209

Change-Id: Ia67a93cd219d2b906b6e3bdd04cc355eeffe857e