Track wireless charger detector timeout explicitly.
Previously we relied on having a continue stream of sensor
events from which to detect whether the device is moving or
at rest.  However, if the sensor HAL is broken in some way
then we might not receive enough sensor events to actually
finish the detection process.  When this happens, we'll
just sit there holding a wakelock indefinitely.

Instead of relying on the sensor event stream, post a delayed
message to explicitly finish detection.

Bug: 10769163
Change-Id: Ia2ed66fe5e7c41a8832df76da9104c13554e1398