Check for presence of carrier app in data notification.
When ConnectivityService checks the mobile data connection and notices that
there is no connectivity, first do a check to see if a valid carrier app for
the current MCC/MNC exists. If so, launch that app instead of checking for a
provisioning URL or hoping for a DNS redirect in the browser. If no such app
exists, continue on as normal.

Bug: 9622645
Change-Id: Ic88b06250af89825d8eee7e4945fb8d7c28d16e1