Always go through bouncer when unlocking
Previously, we skipped the bouncer when the user didn't had a
security method set. Users without a security method couldn't unlock
their SIM cards though. This change moves the responsibility to
decide what to do to bouncer, so SIM pin is still shown with users
without secrity methods. The bouncer will still be skipped if no
authentication is required.

Bug: 13635952
Change-Id: I0b43628c37d769515e97c29fc5fb5337fe46526f