Fix path interpretation on curves
Mostly about the shorthand for quadratic / cubic curve.

We now comply with the spec, which said the behavior should depend on the
previous command.
At the same time, make sure our current point and current control point info
are always in absolute coordinates, not a mix of absolute and relative.
CTS test will be added to monitor this behavior.

Note: Chrome / Firefox are following this rule, while Linux OS / Batik are not.

bug:13913405

Change-Id: I9f8168ed97f87c754c79588984f1f199507b0377