MediaHTTPConnection: support header android-allow-cross-domain-redirect
If present and set to false, media http redirects across domains
will not be followed.  As long as domains are identical, redirects
across protocols or ports will still be followed.

Also fail more seriously if redirection fails or is not supported,
so that media client does not keep retrying the connection.

Bug: 12573548
Change-Id: Ifd2539ad3a90f669d43bd0e82845dbc8ae0b4a3e