Enable display list updates with zero delta time
It's better not to cancel display lists updates. In the emulator, if we use
low precision clock source like refined-jiffies, the display list which
should be updated with a new frame, won't be updated due to the delta time
is 0 . And this makes all the hardware accelerated GUI out of order.

Change-Id: I41dba8965c96b6ad1b6d326ad1d6fa4a0a49173f
Signed-off-by: Tina Zhang <tina.zhang@intel.com>