Add content insets to Toolbar
Content insets allow a Toolbar to define horizontal gridlines for the
start/end or left/right edges of major content (not counting the
navigation button or menu). This allows apps to align their layout
content above or below a Toolbar with the title and other contents of
the bar by using app-defined measurements.

Change-Id: Ibd7e0810bf81070eb5f59a42ad3257f35bc0156a