Layout keyguard according to flags.
Keyguard layout was being skipped because the previous policy does
not layout the status bar. This change lays out the status bar when
it is serving as the keyguard. This change then allows the keyguard
insets to be resized to make room for the IME when password security
is selected.

Fixes bug 14548156.

Change-Id: Ice4f1072b463a22ec763a0cde69f342206bdd15a