Fix PanelView over expanding behavior.
Fixes a bug that the panel height was wrongly calculated and lagging
one frame behind. Also fixes the animation when overscrolling and
then flinging the panel to collapse. In addition, the logic to handle
the over expanding is much cleaner and calculated in an absolut
manner (before, it was relative an really complicated to understand).

Bug: 14487435
Change-Id: If8dbb3e063ef63f51f6dac0ae5bf276480514103