camera2: Remove stray </p> in CameraDevice#createCaptureSession
This was causing strange indentation.
The change also clarifies how to get the supported output sizes in the
ImageReader flow: Although the doc (correctly) instructed the developer to use
StreamConfigurationMap#getOutputSizes(int) with an ImageReader, this instruction
appeared after the bulleted list of output types. Since all other items in that
list specify that getOutputSizes(Class<?>) should be used, the reader shouldn't
be expected to read beyond the list only in this specific case.

Bug: 15574400
Change-Id: Id952b46ae52322d75abbcaffea678d5ba50b22a5