Initial RDNSS tracking implementation.
Add a DnsServerRepository to NetlinkTracker that keeps track of
IPv6 DNS servers received via RDNSS. It supports expiring
existing DNS servers when their lifetimes go below zero and
keeping track of more than just the 2 or 3 DNS servers that are
currently in use, so that if they all expire DNS will continue to
work.

It does not yet expire DNS servers using timers, only when a new
update comes in.

Bug: 9180552
Change-Id: I455699076198f43570a3b0b8ec7e5967514d6086