Finalizing lock-to-app UX
Simplifications have happened to remove the enter/exit layouts
and just use a toast to notify the user, and the in text icons
have been removed.

The enter dialog now has a checkbox to see if the lockscreen
should be locked on exit (since it isn't in settings anymore).

Also updated strings.

Bug: 15836599
Change-Id: I3ab685bf3672c8a7f9d83d12e81f0099968cdfde