Page size not always adjusted
A user can select the pages to print by selecting a range from the
UI after which the selected pages appear in the preview list. Now
the user can further prune pages by clicking on them. If the number
of pages change to two or one the page position and size were not
adjusted. Now after the constraints of desected by clicking pages
are propafated the page size position is also computed.

bug:16297811

Change-Id: I62c0ae8ab5b1d91b120cbf9faa9653ef4ccbd97f