Fix a crash in RemoteViews.
We are now using the application info of the target app
of the remote views to create a deicated context for
loading resources. The code that is looking up the app
info was using the appilcation context of the current
app which may be null. Now we are using the base context
which is never null.

bug:17005860

Change-Id: Ibcf8b6f81fe24d23a82327870741146c8864ce8f