Handle user changes off the main thread.
The remote print spooler can only be accessed off the main thread
by design as calls into it may block for a short amoun of time since
the frist call into the spooler may have to wait for the system to
bind to the spooler service. A recent change introduced a regression
where the user state changes are now dispatched on the main thread.
This change schedules the user change handling to a background handler
thread.

bug:17396682

Change-Id: I059be8f31ba41122cb2967d8afadd19e6b5a08e5