Possible fix for dead handler in Music button event sending
I'm still attempting to root cause, but this is a potential fix. I
suspect the callback is getting registerd on a looper that isn't the
main looper which is then getting killed at some point. This change
will make all legacy callbacks happen on the main looper, which should
prevent the state.

bug:17420281
Change-Id: Ia78e694b32b24398d3c09f49d12632755ef45597