Ensure that the screen on blocker is applied while dozing.
The screen on blocker is used to keep the screen blank while the
system is drawing new content to prevent the user from seeing a
flash of stale content while the screen is being turned on.
This patch ensures that the screen on blocker functionality
is also applied while dozing.

Bug: 17516245
Change-Id: I77c2d0f2b99476a59ad212099f44c63aa2ef9c34