Defer spelling correction with apostrophe
When typing a contraction (such as "doesn't") we don't want a spell
right after the apostrophe, as this will create a false temporary red
underline, and the span split also breaks a kern pair causing text to
shift. This patch detects the case where the cursor is immediately after
such a word and suppresses correction in that case.

Bug: 17641350

Change-Id: I4d09576a31df551c96f820242fd2cbc675506dae