Restored widget provider id not properly updated.
During restore first the provider records and then the applications
are restored. Therefore, we keep the restored provider records in a
zombie state and once we get the package install broadcasts from the
package manager we update the zombie providers to be real ones. The
bug was that while doing this we were not updating the provider id
from one with an undefined UID to the one with the UID under which
its package now runs.

bug:17869323

Change-Id: I20b7cbfd2f90e2f5bf3a6b858d8da99cc1c82841