Fix a wonderful copy/paste time bomb.
While trying to move projects over to libc++, I came across a link
error where these tests were failing to link when built with mma, but
not with mm. When building with mma, they would fail to link because
of symbols missing from the NDK's stlport. Huh?

When cleaning up this makefile a while ago, I removed the redundant
static_libraries := gtest/stlport, but missed removing the
LOCAL_STATIC_LIBRARIES :+ $(static_libraries).  The Android.mk in
bootable/recovery/tests was being loaded before this one, and _did_ in
fact define static_libraries. It defined it to, redundantly, include
the stlport flavored gtest libraries (another wonderful piece of
copy/paste if I do say so myself). Because of this, running mma with
this project set to use libc++ meant that the final binary was linked
with not just libc++ and the libc++ gtest libraries, but also the
stlport gtest libraries and the NDK's stlport.

Change-Id: Ibc545f95f701126ec9f0d0d0ac3d2a0feba78f8a