Let scorers fine-tune scoring heuristics.
Documents Byte#MIN_VALUE as being equivalent to a null score. This
enables scorers to pick a threshold below which a network won't be
used, by setting the score at any RSSI below that threshold to this
value.

Also adds an "RSSI boost" for the active network, so that we avoid
unnecessary switches between two closely-scored networks due to small
fluctuations in signal strength.

Bug: 15432594
Change-Id: I7a8f5f68ef074827d4b1cfbbed0841448498f179