Reset the task stack instead of the view when Recents is hidden. (Bug. 18436876, Bug. 18394126, Bug. 18432310)
This will remove the need for having all the null checks in TaskStackView.
We also remove the check when launching tasks (the fail case is worse
when the activity fails to load).  And the CL also fix the header focus
animation, and an issue with tasks not getting laid out when the stack is
set again.

Change-Id: I9b4cd50d76f72ba3d16dd1b8e2d9bbf9f1f24d2b