Issues with messaging apps on non-primary user account
Bug 18399514

Adjust the contentUri to contain the caller's userId so that when
the phone process tries to open the Uri, it will open the provider
on the correct user.
Also make sure the Uri grants are properly qualified. We only need to grant
permission for sending. Receiving an MMS is always done by the primary
user and doesn't need special permissions. Move various permission grants
from the SmsManager to here.

Change-Id: Ib192f651ab05db9f07e9e6245bb343ed7a55b18e