Notify accessibility for window changes after an app animation end.
Accessibility layer keeps track of the introspectable windows. These
windows are received from the window manager which computes them after
an interesting window transition. The window manager was not sending
the windows to the accessibiltiy manager after an app animation is
completed and as a result the window location reported to accessibility
service was wrong which also resulted in wrong visible to user state
for the nodes in the window.

bug:18517058

Change-Id: I21d65a4e0c0dff9474f7cc47ea819ace5ac1e465