DO NOT MERGE. Don't use visibility flags to determine whether to dismiss Recents. (Bug 18802531)
We can use our normal visibility check to dismiss Recents when the screen is off,
since the system broadcast can occur after the activity is stopped.  We should use
the same mechanism we use to test visibility when launching Recents and just see
if it is the top most activity.

Change-Id: Ib9c01e78fd9221c4fb0ffcc80a01a0c58fb96836