Wipe FRP partition if OEM unlock enabled
Not all devices invoke recovery on every userdata wipe,
so we can't rely on code in the
recovery OS to do this. This results in fastboot -w
not properly wiping the FRP partition. This patch
fixes the issue by having the framework level service
check the OEM unlock enabled bit, and wiping the partition
if it is set.

Bug: 18644051
Change-Id: Id97a29916fe39561700912a920c5741109842bdb