Add capability callback for FLP HAL.
Let HAL implementation tell if geofencing/batching is
supported and which technologies (GNNS, wifi, etc)
can be used.

Still todo: Add ability for GmsCore geofencing to
tell which technologies are supported (instead of
just using it to update monitoring).  This requires
SystemApi change + approval so will do in separate CL.

Note that the classes in the lib are not copied
directly into GmsCore.  The instance will always
be whatever is in the platform.  This is why
the callback is backwards compatible as long as
their is a default implementation (but not if
it's abstract).

Change-Id: I7d6adeb049b89935bc4443785df5d7ef4c730e5d