Introduces navigation keycodes.
These four new navigation keycodes are intended to be used to
navigate a tree hierarchy or list of items.  They should be used to
allow the user to advance forward or backward within a list of
sibling items, expand the currently selected item to navigate the
item's children, or back out from the current item to its parent
item or state.

Change-Id: Ie2b85b3e4c682b0698599d7bcf64a53f6a52d776