docs: M Runtime Permissions overview.
Based on the Google Doc which has gone through tech review. See
first comment for doc stage location.

NOTE: Because the M Preview reference docs are not going to be live on
DAC (they'll be downloadable in a zip archive, I believe), we can't
put in live hyperlinks to new-for-M classes/fields/ methods. Those
have to just be in <code> font with no link. This is consistent across
all the preview docs, and will of course change when these docs are
converted for the MNC release.

bug: 21411602
bug: 21411604

Change-Id: I1163abc1dc884cf27b0e94d266f3979f131db2ab