Fixed a bug with HUNS and dreams
When a notification came in dreamMode and would become a HUN
we were firing its fullscreen intent imediatelly. If that notification
got updated again a moment later, it would also show an additional
HUN even though we just launched the fullscreen intent. This is very
troublesome to handle on the app level, as the notification state
is the same. We are now introducing a cooldown for HUNs when it just
launched a fullscreen intent.

Bug: 19377091
Change-Id: Ib32341c8983f0e977354432ea8d8e98909a13829