Occasionally retest for captive portals once detected
Once a captive portal has been detected, retest for the presence of
the captive portal every 10 minutes.  Some WiFi routers present a
captive portal when the backhaul goes away; with this change if the
backhaul returns and the captive portal goes away, the device will
detect this after a period and remove the notification.

Bug: 21545788
Change-Id: Icabf0a3fff37bccc95f99eba52af13f7a2ddc2f3