Don't duplicate recent tasks of the same affinity.
In change 510e554283d3dc250f836cd3e4abc36d87319333 we stated
limiting the number of document activity entries in recents
by comparing the realActivity of the tasks. This compared was
done for all tasks not just document tasks there by breaking
the use of task affinity to remove tasks with the same affinity
from recent task list. This change limits the compare of
realActivity to just document tasks.

Bug: 22564474
Bug: 18642190
Change-Id: Idc21ec311f194ab2005cb22380ca56303ae0ad94