docs: Clarified that app is running on mobile device, not on Auto
Per user feedback--at times it was unclear whether app was running
on Auto or on a mobile device connected to Auto. Clarified that it's
always/only about an app on a phone or tablet that's plugged in to
Auto.

Also fixed a broken link while I had it open.

See first comment for doc stage location.

bug: 20827220
Change-Id: I52f38674fbdadb6b244af92b66e841b795340fec