Optimize the condition to show IME switcher.
This CL logically reverts Id44724a2a46190382c283a52ece8edffb740807d.

The primary use case of above change was for phones and tablets
connected with Bluetooth keyboards, where temporarily switching back to
'no-physical keyboard mode' might be a bit cumbersome unless we have an
escape hatch like "Show input method" toggle on the IME switcher dialog.

However, such an escape hatch is not necessary for well-designed devices
where we can simply assume that if the physical keyboard is connected
then the user always wants to used it, and if not then not.

Bug: 25432652
Change-Id: I99c19e9f1081cfd3ac2bbda4b829beab4acb7ea8