Consider stack invisible if it got no running apps.
When we close an freeform app by X button while an app is docked on
the other side, ActivityStack.ensureActivitiesVisibleLocked() failed
to make home stack visible, because it's under the freeform stack
which only contains one finishing activity.

We usually will not do another round of ensureActivitiesVisibleLocked
when the activity is destroyed. Need to make sure visibility is adjusted
to the stack below when we start finishing the last app in a stack.

bug: 25668050
Change-Id: Id2a23cb9480221a4e55ba3be9eab51f6efd8a700