docs: Quick fixes to the Google Play distribute docs.
Got some post-submit comments on the Google Play distribute docs.
This CL is for the easy ones that we can submit right away (page
tags and a typo fix), since I can +2 it myself and submit right away.
The larger comments (substantive changes in some material, and a
change of a screenshot) we'll have to leave for a second CL, so
we can confirm the changes.

See first comment for doc stage location.

bug: 25596006
Change-Id: Ib309ca04d243c0ac8b1bcb0d331574f6ff21f510