docs: Fixed typos and formatting in cardflip.jd
Bug was filed for a missing </set> tag in a code snippet; turned out
to be because the angle brackets weren't escaped, so </set> was
interpreted as HTML.

While I had it open, fixed other formatting issues (over-long lines
in a code box, weird indenting of HTML). No semantic changes, just
formatting.

I'll build and stage (see first comment for stage location) but if
it builds and looks good, I'll go ahead and submit.

bug: 25709592
Change-Id: I0ac1e74a230175fcab62021c89884f9eeabb4f97