docs: Fixing normal permission example
The doc previously said that you need just the normal FLASHLIGHT
permission to turn on the flashlight. This was wrong in a couple of
ways: For the (deprecated) Camera API, you need CAMERA permission just
to get the Camera object (which you use to turn on the flash); for
the new Camera2 API, you don't need any permissions at all to turn
on "torch mode". So I've removed this example and replaced with a more
straightforward example of a normal permission (setting the time
zone).

I've filed a separate bug, b/26308110 , to update the FLASHLIGHT
permission docs (once I've confirmed the right update to make).

See first comment for doc stage location, but if the build is
okay I'll just go ahead and submit.

bug: 26104001
Change-Id: Ie0802b4818bd0ba4d3fd026e32620978827e50df