Always show enabled keyboards.
With the latest keyboard layout selection process changes we now only
show layouts that are appropriate for a given keyboard. If a user
selected an inappropriate layout prior to this though, then we need
to still show it in the selection menu so they can remove if it they
desire.

Bug: 25624985
Change-Id: I0707981f5b60c1867954752760f6c780a74507b0