Do not clear exiting flag following explicit remove.
If we begin an exit animation (setting the mExiting flag)
in response to an explicit client remove request, we depend
on mExiting in WindowStateAnimator::finishExit to perform
the final clenaup of the window. If we are transitioning
back in to the app, and it calls removeWindow before the
app transition is ready, we could have handleOpeningApps
clear the mExiting flag leaving this partially disposed window
around due to the above mechanism.

Bug: 26157153
Change-Id: I512fbbd080d423336796284fca490c2c22709180