MediaSession: Define an extra field for staying paused after play
Introduced a new constant EXTRA_PREPARE_ONLY, which allows
MediaController to ask MediaSession to stay paused right after play
requests are called. If this field is set to true in extra bundle and it
is passed during playFromMediaId, playFromSearch, and playFromUri
requests, MediaSession will prepare the given media, but not start
playing it.

Bug: 25841735
Change-Id: I2a2f71a251bd6f475698fe8a5bb696ac70003e2b