Fix: RTL sublocales need to display RTL
When the langage selected in the first step is right-to-left the
localized region names should also be rendered right-to-left,
regardless of the UI language of the system itself.
We only need this to happen for the sub-locale because all labels
are localized in the same language (the parent one), so it is all
consistent, not the mixture we have in the language selection.

Bug: 26278041
Change-Id: I54c4306763458be691a1ba0c2bc2dba613beb8c5