[ContactsContract] Add new API PhoneLookup.CONTACT_ID and DATA_ID
It is a mistake that we put contact id in PhoneLookup._ID.
But for SIP query, the contact id is put in Data.CONTACT_ID instead.
For backward-compability issue, we can't revert the change, but to
save developer from having logic like below, PhoneLookup.CONTACT_ID is
introduced.

if (sip) {
    get contact id in Data.CONTACT_ID column
} else {
    get contact id in PhoneLookup._ID column
}

So for sip query, we put contact id in PhoneLookup.CONTACT_ID.
For other phone lookup query, we put contact id in both PhoneLookup.ID
and PhoneLookup.CONTACT_ID.

BUG:27149017

Change-Id: I4155da29bbacdda0d43164b0ce11c7224a4eee67