docs: Mild fix to a sensors_overview code example
In the previous version, if a device had both a gravity sensor and
an accelerometer, but did not have the Google Inc. version 3 gravity
sensor, we would end up using no sensor at all.  We change this to
have the code use an accelerometer in this case, which more closely
matches the surrounding description text.

Change-Id: I095b94d0d8207d3106db022dc337dd537ee0fa82