Don't run WebView-preparation during an encryption boot phase.
When a device is booting up in an encrypted state WebView won't find the
packages needed for loading WebView and might thus try to use incorrect
webview packages. This in itself might cause us to incorrectly change
the webview provider setting but also causes us to crash the system
server if we try to enabled/disable a package that cannot be seen by the
package manager.

Bug: 27353062
Change-Id: I9349778506e8bec1c56b9b786fe4ed15c7c3260d