Fix APF programs to not generate a failing compare of size 0
Don't generate JNEBS instruction for 0 bytes as it always fails the
ASSERT_FORWARD_IN_PROGRAM(pc + cmp_imm - 1) check where cmp_imm is
the number of bytes to compare.
ApfFilter currently attempts to generate JNEBS for 0 bytes for the
non-lifetime piece between the valid and preferred lifetimes in the
prefix option.
From my testing this fixes RA filtering on bullhead on GoogleGuest.

Bug: 27768019
Change-Id: Id409da55d02767799fb5e3f9d0e9f72ac94d497f