Include animation layer adjustment even while docked.
When adjusting special windows (e.g. docked windows), we omit the
animation layer adjustment. This was largely safe as we place such
special windows above the otherwise highest layer, and we want
docked windows above animating windows. However in the case of
multiple animating docked windows, we need to use the animation
layer adjustment to distinguish the ordering.

Bug: 27712710
Change-Id: Ieb86fd1231d9b04b73de95902832facf4dc0ea3e