Fix slow dismiss animation
An accelerated interpolator is used for dismissing when the
the velocity = 0 which can occur with dismiss all or if the
noti is swiped far enough but not fast enough.

There is no duration specified when the noti is swiped far
enough, and the velocity passed in is 0 so the default
duration is used without taking into account the velocity.

This CL alters the code to pass in the velocity so that a
duration is calculated.

Bug: 27864379
Change-Id: Ifac04bb8d1b714c040102fef035d55737c844157