Not to do handshake verification in SSLSocketFactory
Doing handshake without proper ssl settings will sometimes result in server's rejection of the handshake and thereby the connection.
Here is an example of this kind of problem: https://github.com/ParsePlatform/Parse-SDK-Android/issues/430
Basically the clients (examples listed in the end) of the SSLSocketFactory are going to do more handhshake settings before they call the handshake function of the SSLSocket created by SSLSocketFactory. The
SSLCertificateSocketFactory is doing redundant work on this, and more than that, it may lead to the socket's breaking.

Some of the clients of SSLSocketFactory that set SSLSocket before handshake:
https://android.googlesource.com/platform/libcore/+/09f1b0c/luni/src/main/java/libcore/net/http/HttpConnection.java#204
https://android.googlesource.com/platform/external/okhttp/+/android-5.1.1_r36/okhttp/src/main/java/com/squareup/okhttp/Connection.java#166