Permit package downgrades on debuggable platforms.
This is to make it easier to test on debuggable builds. Debuggable
platform builds trade off being as close as possible in behavior to
non-debuggable builds and being more testable/debuggable. Thus,
debuggable platform builds make no security guarantees and it is thus
acceptable to disable this security mechanism on debuggable platform
builds to help with development/testing/QA.

Bug: 27327503
Change-Id: I19340b95f08c57ff2aba59a08babb6a941c93c3a