Disable fallback webview package before downgrading it.
Now that we support disabling packages before downgrading them we can
do this for the webview fallback mechanism to avoid disabling a fallback
asynchronously (downgrading a package is an asynchronous operation).

We also check that the package is an upgraded app before downgrading it,
to not try to downgrade it unnecessarily.

Bug: 27894939

Change-Id: I541c2fcfe8016e4bcc2d5afff6e393c81e2749a2