Correct PopupWindow spaceAbove calculation.
The existing logic just doesn't make sense for a display frame with
nonzero top. Consider if we were docked on the bottom of a 2560 tall
display and so had say a display frame top of 1280. Then if the anchor
were 80 pixels in to our window it's top in screen would be 1320.
When we would add these two we would end up a value greater than the
entire screen height as the possible space above the anchor. Instead
subtract the display frame, to only consider the possible space above
the anchor, within the display frame.

Bug: 26255254
Change-Id: I95d7575a12a4e319f85c1cf6778970ff37918045