Fix black frame in finishing transition
If either visible or nowVisible is true we need to wait for next
activity to become visible before we destroy the previous activity.

In some code path (eg. clear task top), when starting a new activity,
old activity is first paused and visible set to false with a dummy
transition set. Then finish activity is requested. At this point visible
is already false, but nowVisible is true. We still need to wait for
next app become visible to avoid a black frame shown in between.

bug: 27796252

Change-Id: Ief3d5fc8f11c51a729c424f996ab2597c815e4dd