docs: Moving library docs out of /tools/
The /tools/ top-level directory is going away when we move to devsite
(and the Android Studio docs get their own home) so I'm moving the
non-AS docs out of there. Also adding redirects (and fixing redirects
that pointed to the old locations).

For the time being, I'm also adding these libraries in to the
"API Guides" TOC, just so they're discoverable from somewhere; I'm
expecting this to change post-redesign.

See first comment for doc stage location.

bug: 28270787
Change-Id: I07488251384fbd22d2512a797793cf98654f54e7