Do not update surface for dummy animation as long as transition is set
Transition is not set to ready until it's executed, which could happen
a bit later after the setAppVisibility where we set up the dummy.
It is rare, but an animation pass could still slip in between, and
since the transition is not ready, it modifies the scale and show a
bad wrong frame.

bug: 27742244
Change-Id: I97fb1955e810c7c4f01dc55a28d2e4ce4f47d5eb