Always reload print services
We want to reload print service if anything relevant changed. E.g. this can
be the label. Which label is loaded depends on the the resolve info and
the package manager. Hence short of reverse engineering what fields are
used to determine how this is done it is very hard to say how we figure
out the label.

Hence the only option we have is to reload the print services every time
a print services is added, removed or changed.

Change-Id: Iff5a05dd1cad923da42c360a88559a2cc61c95aa
Fixes: 29765394