Direct Reply: Fix circular reveal origin
Makes sure the circular reveal origin matches the bounds of the
text. This is necessairy because the TextView for the last action
spans the full remaining width and is usually much larger than the
text, which makes the reveal look uncentered.

Also adds an animation for closing the direct reply view.

Bug: 29876991
Change-Id: Ib0ae8a12423f0ecfc573928a39634404c8232e40