Fix VD animator not being properly resumed
VectorDrawable animators remove themselves from RenderThread's
animator list when they detect their target not being in the
DisplayList. When the target got added back into DisplayList
without an invalidation(), the animators were not added to
the running animator list, therefore didn't continue to run.
This CL detects for the case where a VD is back in the DisplayList
and reactivate its animator.

BUG: 30101451
Change-Id: Ib5eb54af954a58a2fa617643bfcf5af40ce708a5