Correctly finish activity in non-focused stack
When activity is finished we first looked for next activity to
show in focused stack. If real next activity to show in place
of finishing one is in the same non-focused stack, we didn't
fully complete the dismissal process and activity was stuck in
FINISHING state.
This CL checks if we're trying to finish visible activity in
paused state and destroy it immediately if top running activity
is visible - same as we do for pinned activities.

Bug: 29458854
Change-Id: I0d5ceb2daa45c0628d89417c8456e132996bcea9