ContextHub: Avoid reference leaks
The ContextHub HAL was leaking references in passOnOsResponse. An
example of this is when a sensor hub reset occurs. If many of these
resets happen, the local reference pool is exhausted and a crash
in the HAL occurs.

Bug: 30706325
Change-Id: Ica5328137a7033f512d7fb1ccf6fa138eab55091
(cherry picked from commit 52dda598963cefc82cfc5e55342ff631a07fd6d4)