ContextHubService: Maintain hub result meaning
We're squeezing a 4-byte signed value to 1-byte.  The naive
(implicit) cast we were using could flip this value from positive
to negative, or vice versa, and completely change the meaning
of the result.

API freeze prevents us from fixing this properly at the moment,
but we use a less naive approach to truncate this value, maintaining
its meaning and logging when we've changed the value.

Bug:30829863
Change-Id: I0e80af9b192066fdf36fee565a4587eb75a5ea7b