Fix RetailDemoServiceTest.
Currently we are using PreloadAppsInstaller initialized with
mock objects. This is creating issues since we are not defining
the behavior of these mock objects. One way to fix this is by
defining the behavior of the mock objects but since we are not
interested in the PreloadAppsInstaller behavior, just create a
dummy PreloadAppsInstaller with no functionality.

Test: Test is passing locally after this patch.
Change-Id: Ic6328303fec99a8eed0aff1b694fa2622e554263