Fix QS CTS failure caused by tile backup
The CTS tests weren't receiving the onTileAdded callbacks since the
shade was not in the listening state when tiles were added. This is
not possible for users (since tiles are added while the UI is
listening), but add a flushMessagesAndUnbind call to make sure the
tiles always receive onTileAdded.

Since now flushMessagesAndUnbind can be called before a real call
to setBindService(true), add a clause that will clear out its
immediate unbind state.

Change-Id: Idd13c432a4af5741813da9a9b1b292a47fd9443b
Fixes: 31147742