Remove  configs in frameworks/base
Remove old configs within frameworks/base. With move to CarrierConfig no
need for the old config_cdma_3waycall_flash_delay resources.

Note that the new carrier_config files replace all values with 300
instead of leaving some as 70 and some as 300. The configs that set the
value to 70 were to fix http://b/17331584, and the configs that set the
value to 300 fixed http://b/24638158. It seems like as long as the value
is above 70ms both bugs should stay resolved.

Bug: 29873049
Change-Id: I15d54e09a4b474b1f83a05cbe777c826cdef368f