Add access tracker to renamed accounts
We are adding access tracker to accounts cached by the
account manager to know when a process saw the account
to white list it for future access. The bug here was
that we didn't return the account from the cache but
the tracker is added when the account is cached. Now
the contract is that inserting an account in the cache
returns the instrumented account which is inteded to
be returned to clients.

bug:28163381
bug:31456742

Change-Id: I5acac1f77a9077b35c5e4892e40d2ab38f3bfc63
(cherry picked from commit b9f84fb6f498644d91e2759a906b00af92cc2f55)