Fix animation glitch with overlapping orientation changes.
Imagine the following scenario:
1. We want to begin a new app transition to an app which will change the
orientation.
2. A rotation animation from a previous rotation change is in progress.
3. Windows have finished drawing the first orientation change,
   so the display is already unfrozen.
In this case, we won't have a chance to select the new orientation which
our pending transition requires, because updateRotationUncheckedLocked
will bail due to the existing animation being in progress. This causes
the app behind to be revealed in the incorrect orientation (sometimes
creating landscape launchers) and the transition to be interrupted
halfway by the screenshot. In this case we can just defer the app
transition until the rotation animation has had a chance to complete.

Bug: 31098404
Change-Id: I1ae1041d3018d5681b46cd39c2d582861f0b014e