Support removal of Icon from preference
This is useful when a preference is used to select an icon (for instance), and the option of setting the selection BACK to 'none' is desired.
Due to logic at line 569, REMOVAL of an Icon from a preference that has one requires (mIconResId = 0) AND (mIcon = null).
(mIcon = null) is not a problem.
This change allows mIconResId to be set to zero. Without this modification, attempting to set (mIconResId = 0) throws an exception on the getDrawable(iconResId) call at line 711 (712 with this mod).