Fixes recents from expanding full size when an app is docked
This is a bug as a result from change ag/1635318 where the
recents/overview area expands to full size when tapped after docking an
app by dragging it to the top. As a result it would expand to full size
and show behind the docked stack.

Fixes: 33335322
Test: ./run-test android.server.cts and ran manual use case from bug
Change-Id: Icf3443631fc3f1c3407d5fd23db077e47ba81110
(cherry picked from commit 299a01fc3cb92115aa1d1f5290198472661fc8e1)