Fixed a bug where notifications could stay expanded
When expanding and locking the screen, the notification
could be stuck in a userlocked state and therefore look
expanded even on the lock screen.
Another case where this could happen was when there was
a race between locking the screen and collapsing a group.

Merged-In: I4f359d727013ee8c9e4174d2855896ba8d3d1a0b
Change-Id: I4f359d727013ee8c9e4174d2855896ba8d3d1a0b
Test: have group on locked shade, expand, click on middle space during expansion
Bug: 33614507