Add getVisualVoicemailPackageName() Make NMR2 voicemail columns public
getVisualVoicemailPackageName() and the IS_OMTP_VOICEMAIL is used by
the voicemail UI to filter out voicemails not from the current visual
voicemail app, for example legacy telephony voicemails or previous
default dialer. These voicmails all represent the same source and are
likely duplicates.

BACKED_UP and RESTORED are used by the system dialer voicemail back
up and restore feature in NMR2. Since there are no API bump, it was
hidden and made public in this CL instead. See ag/1803794

ARCHIVED is used to prevent locally saved voicemails from being
deleted during a sync because it does not exist on the server.

Bug: 34463609
Fixes: 34463609
Bug: 32414216
Test: cts test android.provider.cts.VoicemailContractTest
    android.telephony.cts.VisualVoicemailServiceTest

Change-Id: Iec1e499e763ad92f4a822715b6022e9cbe557c11