Allow shared lib and package name to match
The old package manager implementation was guarding for
shared libs to not match any package name but due to a
bug this was not enfroced and as a result we have apps
where the package and lib match, e.g. GCore. The new
static shared lib support properly enforced the rule
breaking apps. So, this CL relexes the constraint.

Test: CTS test in the topic passes

bug:34598313

Change-Id: I15d22fd1c3a44a27c4b9b8379de3992ae553dd05
(cherry picked from commit 5733d9de2ce946789c4d777179eb251cc3526d27)