Camera2: Do not throw exception if addSurface size doesn't match
During addSurface, the surface size may not match the preconfigured
size, because the surface could be asynchronously changing its size.
Only log a warning in this case, and depend on native implementation
of createCaptureSession or finalizeOutputConfigurations to detect the
size mismatch.

This also reverts commit e56e606f82c94ab9adce1e6b69881653c225cf20,
and enforce the surfaceSize for deferred OutputConfiguration is not
0.

Test: Camera CTS
Bug: 34737377
Bug: 34697112
Change-Id: I4b66166a9a2739a09a16d635d2599a9bcf5c036b