Ensure default and system set bits grandfathered
We added the notion of a default and whether the system set
the setting. This is used for resetting the internal state and we need
to make sure this value is updated for the existing settings, otherwise
we would delete system set settings while they should stay unmodified.

Test: manual

bug:35317326

Change-Id: Iaffde2e7acab53653fd38e669a644e654cc7cd7d
(cherry picked from commit 137015599e666b0a6dc5bfd696025ad61b17d6c7)