Require valid authorities for Uri notifications.
Content changed notifications are really only valid for content://
Uris, which are really only valid when we have a valid ContentProvider
backing them.  This has been implicit for a long time, but we actually
need to start enforcing it based on target API.

We also now tell developers about why their notification requests
are being denied, instead of silently logging.

Test: builds, boots, common operations work
Bug: 34049049
Change-Id: Ie8ab8d8674cff13e3e9269ffddc4ad998cb848c4