Only change bounds if new activity is being added to InTask
When going through the setTaskFromInTask code path, there are multiple
scenarios where an activity isn't actually added to the task.

This change makes it such that the task is only resized if an activity
is actually being added to the task make this more semantically similar
to setTaskFromReuseOrCreateNewTask, where the task is only resized if
it is a newly created task.

This is a cherry-pick of
I111353c76e623925af21a34e395e3d5eb9282e8f

Bug: 35945071
Test: manual - setTaskFromInTask doesn't change bounds if
no new activity is added.
Change-Id: I4b8330ce775ec46812f6095c170a690fdea77736