Handle shared libraries for split apks.
As we dexopt all the code paths of a package, we now append the code
paths to the list of shared libraries. These paths are passed as
relative paths since the apks are staged during installation, and are
not where they will be at runtime.

Part of a multi-project change.

Bug: 34169257
Test: cts-tradefed run singleCommand cts -d --module
CtsAppSecurityHostTestCases -t android.appsecurity.cts.SplitTests

(cherry-picked from commit 832641bb7205126829d425401a5262ab2e26fa80)

Change-Id: I28b1f8db6035ec62a5947679ed09013b1f9523a2