SystemUI: fix incorrect redaction of profile notifications.
There is a race condition for devices with unified profile challenge
when the device unlocks:
1) StatusBar queries TrustManager and caches the locked/unlocked state
   of a managed profile, and
2) TrustManager updates its own record of the locked/unlocked state of
   the managed profile.

If 1) happens before 2), StatusBar will keep an out-of-date state for
the managed profile, resulting in notifications in the profile being
redacted incorrectly.

This CL avoids the race by removing the dependency on TrustManager for
the case of unified challenge.

Bug: 35951989
Test: manual, with logging
Change-Id: I5ce9df017e1de3aca5236d1207fe1867fa5a80c2
(cherry picked from commit c37a22aad6651f1bbcd002a56b35e51db4c3c8a7)