Handle exception in newly spawned thread in the test
Junit catches the throwable that happen in main thread, but not in
custom threads that tests have created and started. As a result,
Errors such as AssertionErrors had led to crashes in the tests.

This CL added error catching in the new thread. The error will then
be thrown in the test thread later on.

Also, transformed the class into JUnit4 tests.

BUG: 64095896
Test: ran core tests
Change-Id: I78a1cceb405d331f6796cf7ee5f7661dc5aec7b4