Catch BadParcelableException in RemoteConnection when setting extras.
Unlike the code in Telecom which can use Bundle.setDefusable(..), this
framework code is used within a Connection Manager's code space, so the
setDefusable method does not work.

To prevent crashing the Connection Manager's app, catching the exception
and ignoring extras in this case (not much else we can do).

I've also filed a bug against the offending component known for putting
bad parcelables into the extras.

Test: Modified code to force this scenario.
Bug: 64499462
Change-Id: If44ad2e2c8e285d513969c6babe5a798242da53d