Fix sync back-off
CL Ifb2fc5207 affected multiple different scenario:
a) When a sync finished successfully
b) When a sync finished with a failure
c) When a sync is canceled by the sync manager

Among those, c) was the actual fix for b/63773598, and the change was supposed
to not change the actual behavior of a) and b).

However it turned out it was breaking b) by interfering back-off scheduling.

This change will revert Ifb2fc5207 for a) and b) to restore the original
behavior with keeping the fix fo C.

Bug: 63773598
Bug: 64597061
Test: Manual test with test app.
Change-Id: I34e5945a7a67d55cd6138f758e97d724d4b9d4b5
(cherry picked from commit 9243d9ca063d9ceb3d8a71f242e1e2facd122383)