Change battery icon to the side of bt icon
Also add a new parameter to control the size of battery icon at
the side.

In spec the height of battery icon should be 55% of the bt icon.
Since bt icon has different natural margin in its vector drawable.
This cl adds a padding and then change the ratio to 75%.

Bug: 65397557
Test: RunSettingsLibRoboTests

Change-Id: I442149392edcec2ee23f651232ffe3e204badc38