Fix clip expansion amount
The clipping rectangle needs to be expanded in order to cover the
content drawn by the stroke. Since half the stroke is drawn on one side
of the bounding rectangle and the other half on the other side, we need
to expand the clip by half of the stroke width.

The previous implementation expanded the clipping rectangle by the whole
stroke width, this allowed for a part of the rounded rectangle to draw
outside of the bounding box.

Test: manual - verify smart select still works and that the rounded
rectangle does not draw outside of the bounding box

Change-Id: I8ca80b7b224d1ac2aa7638405b3cf516bd04b39f