lowpan: Use IBinder for comparison instead of interface
This change fixes a bug where the onInterfaceRemoved() callback for
LowpanManager was not working properly and causing crashes. This was
because we were using the ILowpanInterface objects as the key in a map
to for looking up the associated LowpanInterface objects. This doesn't
work because there may be more than one ILowpanInterface object for a
given IBinder---thus subsequent attempts to resolve ILowpanInterface
objects would always come up empty. The solution was to use the
underlying IBinder object as the map key.

(Cherry-picked from commit aa07c47441ae1e37f87248492459bef336b43155)

Bug: b/67718495
Test: manual
Change-Id: I7575743268cf67c6c2c24d8f327ce38d88d354c7