Reuse the BluetoothBatteryDrawable in QS detail
BluetoothBatteryDrawable creates the drawable lazily using
the latest(and most correct) context. So it could get the
correct color from the theme.

Also add iconScale field in BluetoothBatteryDrawable so the size
of the battery icon could be scaled.

Bug: 67377756
Test: Manual - connected to bluetooth headphone and open the qs
page and detail page. Retest it after changing it to dark theme.
Change-Id: Ic6c871985a89a503221755eed4bc2605ab53f630