Fix race between AnimationListener and DragState
Previously AnimationListener refers ValueAnimator via
DragState#mAnimator, however the variable may be null because the
animation is submitted to the animation thread before mAnimator is
updated and we don't have any synchrnoization between the animation
thread and other threads.

This causes NPE in CrossAppDragAndDropTests.

The CL changes AnimationListener so that it uses animator passed by
argument.

Bug: 69575965
Test: CrossAppDragAndDropTests
Change-Id: I454697136aa35c0a9525ab58171347389b483cb1