Tiny fix to bug when statsd should clear data.
Previously, we always sent a broadcast, even after we have exceeded
the memory limit for this config key. We switch the order so that
we drop the data if the limit is exceeded. If greater than 90% of the
way to the limit, we send the broadcast.

We need to find a way to unit-test this behavior.

Test: N/A.
Change-Id: I6ea40b9e34dceb19805d9af24495d72878f787e0