Add RecoverableKeyStoreDb
Adds database for storing recoverable keys. They are indexed by the
uid of the application that created them, and the alias of the key.
This is the same alias that is used to get the key from
AndroidKeyStore. The database stores the wrapped key, and the version
of the platform key that did the wrapping. It also stores information
about when the key was last synced.

This is used to get the status of the key. e.g., if the platform key
id is not the current platform key id, this is now an unsyncable key.
If the last-synced time is not set, this is a valid key but one that
has not yet been synced. etc., etc.

Test: adb shell am instrument -w -e package com.android.server.locksettings.recoverablekeystore com.android.frameworks.servicestests/android.support.test.runner.AndroidJUnitRunner
Change-Id: I4960452abcdd16c9f39a1f166a086a52dd2f05c0