Set ActivityView background to current top task's color
When ActivityView is being resized to a bigger region and an app
in it hasn't received new configuration yet, it will continue
drawing in old smaller size for some time. This results in black
bars being drawn in expanded area.
This CL tracks the topmost task launched on the ActivityView and
uses its background color to fill this expanded area.

Bug: 72220802
Test: Launch activity on ActivityView, resize and observe background
Change-Id: Id7090a1ad5ec49a31c19fe185fb3815b1788e77f