Updating the text classifier model path
The lib2 implementation of libtextclassifier is incompatible with models
for lib1. To avoid trying to load a lib1 model and failing (which happens
when devices are upgraded from O to P), we need a new update model path.
Using this opportunity to remove smartselection from the filename, which
isn't appropriate any more given what the model is used for.

Test: Ran core framework tests
Change-Id: I79a80d10d920019f5091fe9884f370149d39fe88