[QS] Add header tooltip for long press
Added some hooks in animator to pass along expanded/animating/collapsed
state in a more clear manner. Added tooltip view with the teaser text to
animate in at the top of the view (and fade out either when we start
animating or automatically after 5 seconds).

Added prefs for tracking how often tooltip is shown/hiding it based on
that number.

Updated dimensions for correct spacing (based on redlines).

Test: Visually
Bug: 72528203

Change-Id: I70e4654ed95057fac6d8dbb890d575c2a5d9f215