Fixes crash on navbar using 3rd party launchers
When using a 3rd party launcher, the touch service
connection from pixel launcher and system ui is
disconnecting causing the nav bar to crash when
tapping it because onTouchIntercept returns true
from a NavigationBarView and handles the
onTouchEvent that does not check for a valid
connection. Check for a valid connection to
launcher service in onTouchEvent and before
proxying data to prevent the crash.

Fixes: 74726495
Test: change to 3rd party launcher press nav bar

Change-Id: I6f7efe78599b0829fd90aa0436463e77ec51f634