No divider between floating toolbar item groups
The CL removes the vertical divider that used to exist between adjacent
menu item groups in the floating toolbar, as well as the extra padding
between these, in order to adapt to the new UX requirements. The CL also
centers the text view inside a button, when there is no icon shown at
the left of the text view. This is only relevant when the minimum size
of the button is larger than the text measured width, and we want the
text to be centered inside the button in this case.

Bug: 74032743
Test: manual testing
Change-Id: I309c729eb842d9388066bfb43eb18f33dbfe10b8