Load fallback categories earlier
At the end of the package manager constructor, we are prepare the data
directories for all installed applications. This creates ApplicationInfo
objects which needs to load categories from the fallback categories provider.
If we're populating the fallback categories during this time, there's a
subtle timing issue where we may return a 'null' category instead of an
Integer.

Change-Id: I974a39029d6c15d3986571e610eb0cd106fdc05c
Fixes: 78128442
Test: Manual. System boots and no crash