Add full unlock gesture to FalsingManager
When FalsingManager.onBouncerShown() is called, this will cause the
MotionEvents to no longer be added to the HumanInteractionClassifier,
which means that the classifier will have to work with only partial
gesture, and will not be able to make a proper decision on unlock. This
would cause many valid unlock gestures to fail.

Instead, record the state of the bouncer visibility at the beginning of
each MotionEvent (on ACTION_DOWN). Then, if the bouncer was not visible,
add the full gesture to the HumanInteractionClassifier.

Test: recorded a valid unlock event with inputstudio, then replayed it
with and without the change. Consistently succeeded with unlock with
this patch, and failed many times (but not always) before the change.
Bug: 71762354

Change-Id: Ib93f21bff19c36aa296ddf6ed9a0a3e879eeb69c