Removed IAutofillManager.updateOrRestartSession() logic.
On Android O, when an autofill service returned a null FillResponse the
app session would be done, unless a manual autofill request was made (and this
method was used to implement that logic).

But on Android P the workflow was improved so that when the service returns
a null FillResponse, the session remains in a "semi-dead" state where a new
server-side session is created whenever a view that hasn't been "seen" by the server yet is focused, and this new workflow makes the old logic obsolete.

Test: atest CtsAutoFillServiceTestCases

Bug: 73536867
Bug: 70046972

Change-Id: I9c819ba97b2a9a9de70e60fa9ebe433fb3b7fca8