Delay brightness change
During a fingerprint wake-up, the display state changes bumping up
the display brightness. This means that the screen will flash
bright if the scrims couldn't draw and blank the screen yet.

The fix updates the state of the keyguard, waits for the frame to be
pushed to the display (can take up to 2 frames since we're using
triple buffers), when then asks PowerManager to wake-up the device.

Fixes: 80415238
Test: fp unlock with AOD on and off on devices with displays from
      multiple vendors

Change-Id: I752f9bd5ffd754166ffee7dc15960909dd3f1a81