Use String.equals when comparing conflicting providers
Prior to this change, we were comparing the package name of the
conflicting component with the soon-to-be installed
PackageParser.Package package name using object equality, as used
elsewhere in packagemanager components. However, ComponentName.mPackage
does not use an interned string for performance reasons, so we need to
use .equals here.

Change-Id: Ibfed7cd47f18d06b2f65c6d8f6c46b752859375c
Fixes: 115956948
Bug: 109941548
Test: run pts -m PtsTimeZoneTestCases passes