services: policy: Kill off the legacy dialog
If the systemui is down for the count, changes are that there are bigger problems
to worry about. If we're paying attention, reviewing things like we're suppose to
the SystemUI should never be down.

Also this causes a hiccup in that if SystemUI is restarted (which we added the option)
the legacy power menu shows up every time. That's janky but that's not the reason for
removing this. I just don't think we should have two different dialogs. We're commiting
to the new one!

Change-Id: I9642bfe58668edc706bdd3e5e70522eb4b712160