SubscriptionInfo contains int cardId
We rename the old "String cardId" field to "String cardString". The new
int cardId refers to the publically exposed cardId.

Also deprecate the few-arg constructor. Its okay to use in existing
cases which don't care about the extra arguments (certain tests and
transient SubscriptionInfo objects which aren't really used) but should
not be used in general.

Bug: 80097562
Test: manual
Change-Id: I92264b95da64be77367e869f774bc65f8934fd32