Fixes unupdated system app deletion
A recent refactor of delete resulted in an incorrect test for the
existence of a disabled setting for all system apps regardless of
the isUpdated state. This change ensures that we only execute this check
in the case that the caller has not requested delete of the system
version as well.

This change also removes a few redundant variables from
executeDeletePackageLIF and instead encapsulates them in the
DeletePackageAction.

Test: Delete of an unupdated app is successful.
Test: QR-based setup completes (b/120189588)
Change-Id: I6e789f1c04b82580b4b46178cc69dfd95b50b621
Fixes: 120189588