Remove a special test-only mode from IMMS
This CL logically reverts my previous CL [1] that allowed CTS tests to
relax security restriction in InputMethodManagerService (IMMS) so that
MockIme could be used even when its host APK is instaleld as an
instant app.

With my recent change for MockIme [2], MockIme is now a standalone APK
that can always be installed as a full (non-instant) app, even when
the test APK is installed as an instant app.  Note that this is
exactly how IMEs work for instant apps in real use cases. We have
never allowed IMEs in instant APKs to be used in production
environment for obvious security reasons.

 [1]: I4383129fd9e229a849282e874aff5d4eef1f49f8
      f80087cae1227c48159a2a177729b14f4eb32e48
 [2]: I4ddda5261651ee67c4ff70bc1ab1d4aef64e1a52
      ecf533da5af14418ce232db5d44d27517069ad30

Fix: 121097307
Test: atest CtsInputMethodTestCases
Test: make -j cts && cts-tradefed run commandAndExit cts-instant -m CtsInputMethodTestCases
Change-Id: I9a5da660fc5e7e2334f7d029cded9c323f808084