Fix TaskPositioningControllerTests test
TaskPositioningControllerTests test was calling into the JNI method
setInputWindowInfo. JNI methods aren't fully supported when running unit tests so
it was causing the process to crash.  The JNI method was being called
from InputMonitor which is not needed for these tests. Instead, mock out
InputMonitor to avoid calling into the JNI methods.

Change-Id: I805d6e94bf8f02f073726d36324050dc3dcec016
Fixes: 120697761
Test: TaskPositioningControllerTests pass