DO NOT MERGE Add CarrierConfig to overlay incorrect country ISO
The issues happens when CDMA is roaming to GSM, some US operator's mcc/mnc would change to 204/04
whose country belongs to NL not US. It doesn't always change but error occurs once changed.
MMSService would check if locale matches with phoneNumber.
The error results in HTTP header bringing incorrect MDN to server.
In Q, AOSP fixes it by using CarrierConfig to enforce 'US' as iso-country for all error scenarios.
Since the change is great, migrating all the patches to P for the time being is risky.
So, we use the same CarrierConfig as Q and narrow the range to MMS only.
That's why we use "DO NOT MERGE".

Bug: 120448454
Bug: 122332763
Test: Build pass
Change-Id: I74a7be348f37c06c015c50f96fbcf7db4fb15246
(cherry picked from commit f636191beaa35966ec0287b876502a77c160fb7b)