Update configuration to activiy after letterbox.
We didn't push the new configuration if DisplayContent decides not to
rotate the display in regard of activity's new requested orientation, so
make sure we push it.

Note ensureActivityConfiguration() checks activity's state so that we
don't push configurations to stopping or stopped activities. Therefore
we don't need to check that.

Fixes: 122984646
Test: Manual test. atest WmTests:ActivityRecordTests
Change-Id: If3594207f8f50f8239023b35ae69b32999e53a76