Fix crash & positioning of notif-based bubbles
We don't need to be adding the row view until the bubble is expanded so
I moved the code to where we actually show the view & added a check to
remove it from the parent (notif stack scroll layout) if it's attached.

Fixes an issue where the expanded state could be in the wrong position.
This was because the y calculation for the expanded state happened
before it's new height was calculated so it could be incorrect.

Test: manual (1) - Have some notif based bubbles
                 - Dismiss the bubble stack
                 - Receive an update for one of those bubbles
                 => SysUI doesn't crash
             (2) - Have 2 bubbles
                 - Tap on the stack
                 - Tap on the 2nd bubble
                 => Expanded state should be positioned correctly

Fixes: 127383589
Fixes: 126550620
Change-Id: Ic2c20d81a2997665810eb9240b28da90f1a9fad5