docs: Javadoc cleanup
There were a number of Javadoc style problems in the
WifiNetworkSuggestion.Builder.build() docs, which caused some formatting
problems.

One specific problem worth noting: A {{ causes trouble at publish time,
because it looks like the start of a Django directive. But escaping the
{ -- e.g. &#123 -- doesn't work, because the Javadoc compiler helpfully
escapes the *ampersand*, so the HTML output is "&amp;#123", which
renders on the page as "&#123" instead of "{". So putting a space
between the two braces seems like the simplest solution.

No code changes here, just Javadoc comment fixes.

Tested by building the docs and staging to:

http://go/dac-stage/reference/android/net/wifi/WifiNetworkSuggestion.Builder#build

Screenshots (in case the stage gets overwritten):
https://screenshot.googleplex.com/o4p7efFGCbJ.png (1)
https://screenshot.googleplex.com/nkhEqU54LNb.png (2)

Bug: 128536865
Test: make ds-docs (then build & stage)
Change-Id: Ie04bc4dc517443a6899aba989beeb38b9299826e