DO NOT MERGE Turn off LAYER_TYPE_SOFTWARE on BatteryMeterView
We still need to save a layer for the BlendMode.CLEAR paint to do the
right thing. But with this we can save just the layer needed to draw the
battery instead of the whole battery meter view.

Test: visual
Bug: 130536257
Change-Id: I9d5844440250ccd9174e6c001e051c9adf62a2d0
(cherry picked from commit b4c703b8e30e50fa451171be563c556704db5ef4)