Add dumpable to QSTile. Dump QSTileHost, QSPanel, QuickQSPanel
Dumps state of all the tiles in QSTileHost. The dump of
QSTileHost should match that of QSPanel. The dump of
QuickQSPanel should match the first N items of
the QSPanel dump. If there's any discrepancy, it's cause for
investigating.

It also dumps state of QSTileView.

An example of the dump can be found in:
https://paste.googleplex.com/4811551678660608

Test: manual, dumpsys
Fixes: 129544734
Change-Id: I3249ffcac1bc5ab3d7666bc8884bda4b3a4eb4e2