Don't slide down "At a Glance" when notifs aren't visible.
When transitioning from AOD to lock screen, "At a Glance" slides
down to be positioned on top of showing notifications. This looks
bad when there aren't any notifications because "At a Glance"
may overlap with the custom clock face. This change leaves
"At a Glance" where it is when there aren't any notifications to
show.

Bug: 129348218
Test: Added test points to KeyguardClockPositionAlgorithmTest
Test: Manual - AOD to lock screen transition w and w/o notifs
Change-Id: I0fbdb54330b8f0f247940808286b79b779171dc9