Check if DSU is in use before feature flag
We are using a feature flag to protect DSU and the device. But when
a Dynamic System is in use, feature flag is set as false by default.

This CL makes DSU bypass feature flag and start GSI service on boot
completed, check if the device is currently running a Dynamic System,
then, stop GSI service if the device isn't.

Bug: 132575851
Test: restart into DSU, check if the "In Use" notification is shown.
Change-Id: I9dabf9bc1716ef18d513278415c6544a5467d510