Don't adjust task bounds for IME during unminimizing docked stack
In split-screen mode, when home is resizable and IME is showing for it,
and launcher doesn't hide IME before launching next task, the current
stack bounds (minimized bounds) may be used as adjusted bounds for task.

Moreover, if the adjusting of bounds for IME and the moving divider for
unminimizing act at the same time, the animation may look flickering.
So this CL add a condition to not adjust task bounds for IME until the
docked stack is unminimized visually.

Fix: 132451731
Test: 1. Launch an app to primary split screen.
      2. Press the search bar of launcher.
      3. Type a word on IME then click the search result.
      4. Check the primary split screen doesn't have black area.
         (wm.Task shouldn't keep the adjusted override bounds)

Change-Id: Ife087b2aa50f1d689803acd5dc54b153b0e8b24b