Add a mapping for kDownloadMetadataSignatureMismatch.
Add a mapping for kDownloadMetadataSignatureMismatch (26) from
error_code.h. This involves creating a new system API constant, but it's
just a simple definition so we don't have to hard-code '26'. These
constants are already set in error_code.h. A TODO has been added to
create this constant in the next release.

Test: manual
Bug: 132827080
Bug: 133396459
Change-Id: I4d59702879f2ed654dd467a1ae2dc8e400a0938d