Makes current user not clickable
When clicking on the current user, SystemUI will try to switch to the
current user and fail silently. There's no reason to go to that path, as
we could prevent the current user from being clickable.

The only times the current user should be clickable is if it's a Guest
(and we want to remove it).

By making the current user not clickable, TB will not announce it as
such and will not create confusion.

Fixes: 139920678
Test: manual, switching accounts and into Guest
Change-Id: I413f62c9e23249c92b192ed4fb4974d75c1543c5