Revert "Fixed default angle to be Orientation.LEFT_RIGHT"
This reverts commit d31d0967209775ae352092d3125adfd59c8040d0.

Reason for revert: Although this change fixed the behavior change for for creation of
GradientDrawables defined in xml, the default value for GradientDrawables defined
programmatically is different. The default orientation for GradientDrawables defined
in xml is LEFT_RIGHT, however, the default orientation for GradientDrawables defined
programmatically is TOP_BOTTOM. Since a fix for AAPT has been made to automatically
insert an angle measurement of 0 if one is not defined, we can revert this CL and target
a proper fix in master.

Change-Id: Ib8983386832fb25f53b5e68e76e9d41d9d26fec9
Merged-In: Ib8983386832fb25f53b5e68e76e9d41d9d26fec9