Changed car status bar to show or hide by setting layout height instead of removing/adding the view.
Adding/removing the car status bar via window manager causes issues if
trying to do too quickly since the calls are asynchronous. So if the
status bar is under the process of being removed when we want to show
the status bar, it will throw a BadTokenException because not all of the
references have been cleaned up.

Bug: 142272085
Test: manual
Change-Id: Ia0f46ba4f7582aaadd24dd885788609f027fa156