Give INTERACT_ACROSS_PROFILES parity with _USERS for bindServiceAsUser API
INTERACT_ACROSS_PROFILES should work as a replacement for INTERACT_ACROSS_USERS so we don't have to give privileged system components a permission that is more powerful than is necessary.
Javadoc and annotations of the bindServiceAsUser API will be modified in a subsequent CL.
Unlike INTERACT_ACROSS_USERS, calling an API for a user in a different profile group will throw a SecurityException.

BUG: 140807471
BUG: 136249261
Test: atest ContextCrossProfileHostTest
Test: tested locally that granting the permission allows access to
bindServiceAsUser

Change-Id: I2878067bb2369d26a9ac5504ddd8c0e753f67898