Add BlobStoreManager APIs.
Apps can use these APIs to contribute data blobs to the blob
store managed by the system. This would allow other apps which
require the same data blob to access it from the system instead
of re-acquiring the data blob and having another copy of it on
disk. By sharing the data blob with other apps on device, apps
can avoid taking the total blame for the disk space occupied
by the data blob.

Apps contributing the data blobs can do so by creating a
blob store session and writing blob content to this session.
This session can be opened/closed multiple times using it's id
but once all the content has been written, it needs to be
committed to be considered part of the blob store. Once committed,
it is not allowed to make any more changes to this blob.
Before committing the blobs, apps can specify what other apps
the blob can be shared.

[Implementation will be in a follow-up change]

Bug: 143559646
Test: n/a
Change-Id: If5b4cfb618c202a3f01b2d1b486501246c59d058