Tweak the defaults of module stubs
We will have 4 stubs for each framework module:
- stubs-publicapi with check_api
- stubs-systemapi with check_api
- api-module_api with check_api, just module_api
- stubs-module_api without check_api, module_api + priv_apps

This is because we want the stubs of module_api to include the priv_apps
system APIs, but don't want the api .txt files to contain them.

The check_api attrs will be setup later, when the individual modules
have been set up to inherit these defaults correctly.

Bug: 147768409
Test: m
Change-Id: Ie0fbff6949c8df3d6211558de7373bddb8b95768