Notification statsd logs: Visibility changes.
These are currently logged in NotificationRecord.setVisibility().
Rather than add a dependency on the NotificationRecordLogger to
NotificationRecord, I've passed one in to that method.

I considered but rejected doing the same for
NOTIFICATION_INTERRUPTION, which is logged similarly in
NotificationRecord.setInterruptive().  The circumstances under which
that log gets written are challenging to understand, and I'm not
currently convinced that it is worth porting.

Test: atest NotificationManagerServiceTest NotificiationRecordLoggerTest
Bug: 146488473
Change-Id: I956cd4ededa470f91244305d4c26fb981d4c94a4