Refresh list on work profile state change.
When the work profile is enabled or disabled
from anywhere on the device, the share sheet and
the intent picker will refresh to either show
the list of work targets or show the empty state
screen that the work profile is not enabled.

Showing that empty state screen also has a
button to enable the work profile. Pressing
it enables the work profile and also refreshes
the list.

Fixes: 149497248
Test: manually toggled the work profile from quick
settings, from the empty state screen button and
from launcher for both share sheet and intent picker.

Change-Id: I25e38b0cb5824ebdcf6255ed1959e9d7c6fac445