Uses TYPE_ACCESSIBILITY_MAGNIFICATION_OVERLAY for window magnification
Window type, TYPE_APPLICATION_PANEL and TYPE_APPLICATION_PANEL, are not
trusted window type while TYPE_ACCESSIBILITY_MAGNIFICATION_OVERLAY
is a trusted window type.  ag/10493916
The touch event dispathed within the window that is not trusted
window type might be referred as an obscured event that would make
users unable to turn on a11y settings.
So we change the window type for the windows in window magnification.

Bug: 144075853
Test: adb shell settings put secure window_magnification 1
1. Turn on window magnification
2. Go to a11y menu service setting and click on the switch.
3. Move window magnification overlay above the setting dialog.
4. Click the dialog button within the mirror window.

Change-Id: I9cc2cda7e628b4c2208b5eab1ab1986394088c48