Use REPLACE_EXISTING for rename(2) workaround.
We implement a workaround for Os.rename() failing with EXDEV, by using
Files.move() instead. However, since we didn't pass in REPLACE_EXISTING
as an option, Files.move() would fail if the destination file already
existed. Since the semantics of rename(2) are to replace a file if it
already exists, the work-around should do the same; so pass the flag.

Also, log if the workaround throws an exception; that would have made
this easier to debug.

Bug: 151078664
Test: atest com.android.tests.fused.host.FuseDaemonHostTest#testRenameAndReplaceFile
Change-Id: Ie84e89f857c2cc71e3b373db2032f700d066cf6b