Move icon logic out of NotificationEntry
It's been on the list of things to remove for a while and now the
conversations feature has some extra needs that can't be addressed by
the existing system.

Moves all logic into IconManager. Icons are still stored on the
NotificationEntry (boo) but they're all encapsulated in a getIcons()
object. All of this logic should eventually move into the content
inflation pipeline (probably?) but for now at least it's self-contained.

Bug: 112656837
Test: atest SystemUITests
Change-Id: Iecbd8999c2c866c7217ad9d0e090fb1db7ddd690