Remove UnsupportedAppUsage from nativeNewInstance([BII)
Test: make

This method is marked as UnsupportedAppUsage, but that was for the
signature prior to I0bf8c4c72cc118d2799a59e8eb6c2cede546fceb (which has
not been released). This was a false positive, anyway. No app should be
using this while there is an almost identical public method.

Change-Id: Ib71a58be6ec72cd4c52e8fca3c2cd4055df55ecf