Allow BIOMETRIC_CONVENIENCE to register with BiometricService
BiometricService should be able to handle things other than the public
android.hardware.biometrics APIs - use this class as the source of truth
for all biometric sensors, configurations etc. On the same note, just
because a sensor is registered here does not mean it will participate in
the android.hardware.biometrics APIs.

This default implementation allows device with BIOMETRIC_CONVENIENCE to
successfully boot and pass the CTS tests.

Bug: 150014844

Test: Manually configure a device to be BIOMETRIC_CONVENIENCE only,
      then run the CTS tests
Test: atest com.android.server.biometrics

Change-Id: I704914b2529e17582ef9b87498e2d57184c560fc