Do not expand a notification if its guts are exposed.
A notification may be expanded by both ExpandHelper and SwipeHelper if
both the touch slop and long-press thresholds are crossed at the same
time. This is quite difficult to trigger with conventional touches, but
is quite easy to trigger with a deep press.

To prevent this, the SwipeHelper should not receive touches if the
ExpandHelper has expanded a notification (mExpandingNotification), and,
conversely, the ExpandHelper should not receive touches if the
SwipeHelper has exposed a notification (guts != null).

Bug: 148172385
Test: deep press on notifications; guts are exposed without
      visual/haptic glitches
Change-Id: Id81c9034bfa9c29b35d43a6125537ca5f9a2e9c4