Add carrier config to indicate if cep is supported on peer.
Adding new carrier config option to indicate if conference event package
data is supported on the conference peer.  We already have an option for
the conference host.  This new option is used to indicate when the peer
can or cannot support the CEP data.  When not supported, even if the
network provides CEP data, it is ignored by the device.

Test: Manual testing using test intents.
Test: Added unit tests for these cases.
Bug: 150860282
Fixes: 154245604
Merged-In: I47b5e49acf08e085ea23e21a5f6b2d5a5d00bc09
Change-Id: Ib8d8f171699f21f45b0ce2eec2dbe75d7466b76a