Update taskAffinity with application uid.
F365d3a8e, Limit Activity taskAffinity to application uid for
security vulnerability issue. It modifies the taskAffinity during
ActivityRecord created, but the format of taskAffinity is also
compared in shouldUpRecreateTaskLocked. AppCompactActivity uses
this function to judge if it should recreate the task when
navigating up or not.

The CL update taskAffinity format with application uid to align
above changed.

Bug: 153390756
Fixes: 153570741

Test: atest ActivityStackTests ActivityTaskAffinityTests
Change-Id: I6a5c062d4869fbda46ae70096ee1d2be7140efe5