Update display orientation that was interrupted by recents animation
When launching an activity with different orientation, the recents
animation can still be started before the transition is completed.
During animating recents, the display will keep original orientation.
So once the recents animation is finished, it still needs to check
again if the top activity may have pending orientation change.

Fixes: 155863908
Test: atest RecentsAnimationControllerTest# \
      testClearFixedRotationLaunchingAppAfterCleanupAnimation

Change-Id: I0f85dae9f55073fefe366d1f82929f5d250fc866