PipAnimationController: Make sure we set crop when resetting scale
For example if we are going from small to large, then the
crop needs to be expanded in the same frame that we reset the
scale and complete the resize. This is how the free resize
path works, so we just make the animate resize path also
work this way.

Bug: 150810705
Test: Resize pip small to big. Everything works!
Change-Id: I41c8b41bada159908227991001dfa24e70bf8545