Start screenshot timeout even if saving fails
Currently we only start the timeout if the saving task completes
successfully. This means that if it fails, the preview hangs around
indefinitely (though it's still dismissable via the x button).

This change moves the timeout code so that the timer is started when
we get a response from the saving task, regardless of whether it
succeeded or failed.

Bug: 158225178
Fix: 158225178
Test: manual
Change-Id: Ica15a12e3f9abf49d8d7d12d3947d2bfb8d61252