Use rotation animation for orientation change of non-occluded activity
If there will have orientation change on existing visible activity, it
is better to use rotation animation to cover it, otherwise it may look
as a obvious jump cut.

Also clear fixed rotation launching app if it is removed from display
to ensure the fading animation of system bars can finish.

Bug: 159707982
Bug: 159099239
Test: DisplayContentTests#testApplyTopFixedRotationTransform
      DisplayContentTests#testRecentsNotRotatingWithFixedRotation

Change-Id: I01b6ddf97e9d1025685e6f14f0c05f3a9cc95cfa