Prevent dagger from hanging onto DozeService
Evidently, dagger hangs on to classes marked with @Reusable. This is
not desirable for DozeService. Instead, institute a better fix
which is binding an interface implementation into its subcomponent,
which is the part that needs DozeService to exist.

We can't bind DozeService directly as Dagger complains that you can't
bind classes that already exist in a parent scope. Binding an interface
implementation, however, works well.

Fixes: 165208002
Test: manual
Change-Id: Iae061d636b6e4bdc8bc6a03e9c7daf900610d19f