Perform traversal after finish drawing for Blast Sync
Normally, a perform traversal wouldn't get called if
WSA.finishDrawingLocked doesn't return true. It only returns true if
the previous drawState was DRAW_PENDING which may not be the case if
it's not the first draw.

With blast sync, we will get a finish draw but not necessary the first
draw. That means we may never trigger the perform traversal. We always
wait for a perform traversal before notifying blast sync that we're done
in case there's more data to gather into the transaction.

In this change, always trigger perform traversal if finishDrawing was
called because of blast sync.

Test: Hard to trigger, but existing sync works
Change-Id: Id95c0c84418cecd0e73befafe524102f9e3faad2