Disable the services before each test
We set the AutofillService in some tests but we don’t disable
before the next test, the test execute time is very quick, sometimes
it may affect the next test. To reduce the problem, we disable the
AutofillService and AugmentedAutofillService before the test to make
sure the test state is clean.

Bug: 165822715
Test: atest AutofillPerfTests
Change-Id: I77c9cc0749c38823a4305a7a0a23f940295db828