Fix checking of admins in resetPassword
Ensure resetPassword throws the right exception or returns the right
value when called by Device Owner, Profile Owner or Device Admin.

The methods for getting the ActiveAdmin associated with the DO/PO
did not work with a CallerIdentity object that had a null componentName.
Now they can cope with that.

Bug: 171407703
Bug: 171408165
Test: atest com.android.cts.devicepolicy.OrgOwnedProfileOwnerTest com.android.cts.devicepolicy.MixedDeviceOwnerTest com.android.cts.devicepolicy.MixedProfileOwnerTest
Test: atest FrameworksServicesTests:DevicePolicyManagerTest
Test: atest  com.android.cts.devicepolicy.DeviceAdminHostSideTestApi23#testResetPasswordDeprecated com.android.cts.devicepolicy.MixedDeviceOwnerTestApi25#testResetPasswordDeprecated
Change-Id: Ibe5c8587f07f1416ffa72dd6f4a8c20a9ec06115