Apply notification action list bg color to actions container
The @color/notification_action_list_background_color, which is @null in
vanilla android, should be applied to the actions_container_layout so
that the bubble_gone_padding_end, which is also a part of the container,
is also included in the background. Without this, that padding is not
included in the colored background.

Bug: 168637667
Test: build, test color overlay on ARC++.
Change-Id: I23306a130e543593c52231fe8dbc896cbe2874b8