Examining if started activity violates lock task mode earlier
The lock task mode violation was used to check on a given task,
so a new task must be created before knowing whether the activity
can be started in lock task mode. If the activity shouldn't be
started, the activity and the task will be removed then.

Instead, this CL allows the lock task mode violation check to be
perform on the activity which is going to be the root activity
of the task which hasn't be created yet, in order to reduce
unnecessary costs if the activity shouldn't be started.

Bug: 171946327
Test: atest LockTaskControllerTest

Change-Id: Ic79368e5197742c34fdde61781d4a0b048540736