Grant CHANGE_DEVICE_IDLE_TEMP_WHITELIST to call log provider.
The call log provider needs to grant an allow-list exemption in
broadcast options when sending the ACTION_NEW_VOICEMAIL intent to it's
directed recipient.  This is required so that the voicemail app can
start a background service to do some time consuming (but timely) work
in the background upon receiving the intent.

Test: Manual test with Call log provider.
Bug: 173155475
Change-Id: I3e84d65292ec74ffe374fc997e57c4c688f71b9d
Merged-In: I3e84d65292ec74ffe374fc997e57c4c688f71b9d