Add System APIs for TimeZoneProvider
First step of moving LocationTimeZoneProvider from the
com.android.location.provider API surface to system API.

This copies / renames various APIs, generally dropping "Location" from
the name to try to generalize them. The
com/android/server/location/timezone classes have been updated
to use the new classes but the names have been left for now as they are
an internal detail and it's not clear which would be "location specific"
and which might be general.

Bug: 175633818
Test: build / treehugger
Change-Id: Ied2b43336ef80d2b4ed2eba8849ea1976145df8d