Fix fingerprint dialog color contrast for dark theme
The code to set the color of the help text shown on the fingerprint
dialog does not currently check the context theme. This means that the
wrong color is used when the system is using a dark theme.

This commit fixes the issue by ensuring the appropriate themed color is
used. It also increases the contrast of the dark theme color so that
the text is more visible against the dialog background.

Test: Set system theme to dark mode and trigger the fingerprint dialog.

Bug: 174831630
Change-Id: I41dfb536e72063ae8ea84d51fa0e50613f920090