Move launcher activities to the back when back pressed on root task
This change partially reverts the back button behavior change introduced
in: I48ec35f841ab3b306fe80845150000c390908f5e

Instead of always moving the task to the back when the back button is
pressed on the root task, we only do it when the activity that is open
is a "launcher activity". These are activities that act as the main
entry points into an application, indicated with an intent filter with
ACTION_MAIN and CATEGORY_LAUNCHER.

Bug: 176235889
Test: atest WmTests
Test: atest CtsWindowManagerDeviceTestCases
Test: manual: flash crosshatch, press back button on a launcher activity,
observe task is moved to back; press back on non-launcher activity,
observe task is finished.

Change-Id: I1cbf85f1a007aeedcb66b9343d182188bacc78e7