Add onFrame methods to face auth/enroll client monitors
Updates the AIDL face sensor implementation to forward the
onAuthenticationFrame and onEnrollmentFrame events to appropriate
methods in the FaceAuthenticationClient and FaceEnrollClient monitors.

This implementation also converts the hardware AIDL models
AuthenticationFrame and EnrollmentFrame to equivalent framework classes
FaceAuthenticationFrame and FaceEnrollFrame, respectively. This will
allow the framework to define AIDL interface methods with these models.

Test: atest CtsBiometricsTestCases

Bug: 178414967

Change-Id: I093c4396dddab9cb9fb8fd8ec3426b4f3a81abeb