Remove unnecessary permission check.
There's no reason for getPowerSaveModeTrigger to require the POWER_SAVER
permission. The API is only marked as a SystemApi because it was designed
for a specific use case and 3rd party apps don't need to know the
information, but there's no problem if they get the value. All the value
tells you is whether the device will turn on battery saver automatically
based on the percentage level or if it'll turn on automatically based on
a different dynamic signal.

Bug: 174241855
Test: atest android.os.cts.PowerManagerTest
Test: atest android.permission.cts.PowerManagerServicePermissionTest
Test: atest CtsPermission2TestCases
Test: atest SettingsProviderTest
Change-Id: I3146a681d9c5da9c68a7e6b38b50314b293dc924