Update app links resolution to handle no category Intents
If an Intent has no categories except DEFAULT, and there is only 1
non-browser result resolved along with 1+ browsers, the set of browsers
will be ignored and app links will be run as if the intent was resolved
with the BROWSABLE category.

This means callers who omit BROWSABLE but resolve to only 1 app +
browsers will automatically launch into the app.

Bug: 181570854

Test: manual, am start launch test intents

Change-Id: I1adfd41c64abd46dc98930ba36409550955275a9