Relax null checks to fix tests.
The nullability annotations on these methods are aspirational,
and a recent CL started routing many existing call-sites through
them, so we'll need to start allowing nulls.  I'm not at a full
workstation right now, so this CL simply relaxes the enforcement,
and a future CL will adjust the API annotations.

Bug: 182091763
Change-Id: I43d49747deb8c7881d942eba48237eb0fccc7acd