Fix FileSystemProvider open for external USB files
Ie08b94cf79c86e9f87639e4fbdd82b2732771187 introduced a regression
where we try to coerce *all* files to a media store URI to open with
the correct media_capabilities for the caller.

Unfortunately, external USB files cannot be coerced to a media store
URI and transcoding is not supported there anyways.

Now, we avoid coercing USB files, additionally, if the coercion fails
for any reason, we fallback to opening via the filesystem directly.

Test: Manual
Bug: 180802673
Change-Id: Iba8505669897d4ebfef5232707b8e2fec9bed76f