Rename #updateInputMethodWindowStatus() to #setDismissImeOnBackKeyPressed()
This is a purely mechanical method renaming.  There should be no
behavior chagne.

To summarize,

  WindowManagerInternal#updateInputMethodWindowStatus()

should be renamed to

  WindowManagerInternal#setDismissImeOnBackKeyPressed()

and this CL actually does so with also removing unused parameters.

The method in question was originally added by me for the
true-IME-target-window project [1] at Bug 34628091 then updated to
propagate one more boolean value from InputMethodManagerService to
PhoneWindowManager to address a feature request from Wear team [2] at
Bug 33824860.

After that, we also introduced another method

   WindowManagerInternal#updateInputMethodTargetWindow()

for the true-IME-target-window project [3] and that's what we ended up
having relied on.  Given that #updateInputMethodWindowStatus() is used
only for Bug 33824860 to propagate "dismissImeOnBackKeyPressed"
parameter from IMMS to PhoneWindowManager, it'd make more sense to
rename its method name as well.

 [1]: I54fb1faf513883a54293a756e4c9d2ae9453d778
      69e68024220d3f18b08cbb314c7991f8f76cbe49
 [2]: I20721547c73360a70b5fc5cbe06824d577d1768a
      d6475a682d9651a651f60856baef9b17b4633b13
 [3]: I032b91ce722a16b17518a5d88015c574d7d3e51b
      99e1f6e04b63a4d85f42fda292001e8b77060093

Bug: 171827834
Test: presubmit
Change-Id: I3b32d43615453d94d5df1565fe1ce81dc08526de