Ensure Display Settings Overrides Do Not Impact TestDisplayContent
Display overrides such as width, height, and density
should not override values set in TestDisplayContent.
Otherwise, test conditions are not consistent and can
lead to test failures on some devices.

To prevent this, set the uniqueId of the DisplayInfo
of the TestDisplayContent to a value different from
that of the actual device displays. This ensures that
DisplayWindowSettingsProvider does not apply override
settings by comparing against uniqueId.

Bug: b/182499175
Test: atest DisplayContentTests
Change-Id: I0ad10c2b3dc30849373a938bfd06533731f1a6dc