Move Call state into Call.Details
The Telecom Call object is primarily a control object which allows
control of a call and receipt of events related to a call.  The actual
properties and data pertaining to a call all exists in Call.Details...
except for the call state (RINGING,ACTIVE, HOLD, etc).
This CL moves the call state to the Call.Details class to remove this
one dangling property of the base Call class.

Test: Run Telecom CTS test suite.
Bug: 163085177
Change-Id: I59413d3e14d03b48263a5cb351fe4e9a553b578e