Bump Standby Bucket for FGS sooner.
Instead of waiting 30 minutes to update the standby bucket for FGS, we
will now only wait for 1 minute due to the new restrictions.
Additionally, the continuity check for interaction is now reduced to
ten minutes from two hours. This is to ensure there are no irregular
bucket de-elevations from ACTIVE.
This change is gated by targetSdkVersion check and is only enabled if
apps target S or above.

Bug: 183972877
Test: atest OomAdjusterTests
Change-Id: Icb2638ba5a20ebb79d55b0575bf351eadc11f8f6