Fix issue with acquiring singleton content provider from another user
The content provider host app could be living in a different user from
the client app's expectation when it's singleton, this could result in
client app ignores notification of the start of the content provider
because of the discrepancy of user id. Now remember the original
expected user id so client app can clearly know whether the incoming
notification of the start of a content provider is what it expects or
not.

Bug: 181789697
Test: atest CtsContentTestCases:android.content.cts
Test: atest FrameworksCoreTests:android.content
Test: Manual - steps in b/181789697#comment1
Change-Id: Ic909620059e1f989a5e7c8cf3ede2b8d63814a96