Reset the features in A11yInputFilter per display
The A11yInputFilter is an input filter for implementing many
accessibility features on multi-display. When any display is
added or removed, all features of all displays would be reset
now.

To avoid server side bad user experience, we should only reset
the features and the event stream state based on the changed
display which is added or removed.

Bug: 187996964
Test: a11y CTS & unit tests
Change-Id: I39ae59e12690240adbb0e81338dcb4c736a92197