Fix issue when adding existing connections via remote connection services.
When an existing connection is added to represent an IMS conference
participant, it is added with an associated parent conference specified.
This was NOT passed along to the connection manager via the
RemoteConnectionService API since there is no equivalent API present for
that operation.  To work around this, we stash the parent conference's
ID into the connection extras.  We will use that later so that Telecom
can know which conference is the parent of this call.

Also, removing restriction in RemoteConnectionService which would ignore
conferences with no children.  This assumption was incorrect for VOLTE
conferences since they will start with no children.  As a consequence we
would ALWAYS skip adding IMS conferences to the connection manager, which
would mean that it had no way of knowing about the conference that the
existing connections are associated with.

Test: Manual test with connection manager APIS on live network; make
conference and verify that the correct objects are being created wrapped
by the connection manager.
Bug: 188420526
Change-Id: Ie58afed7a3b7eeaa7e329e80479d273e4c50ec82

Change-Id: I4250f9459c7a1b82936583a10e93d049fdfb4c5d