Ensure backlight doesn't change when rbc is toggled
This change fixes 2 issues.
1) the delay in autobrightness updating when rbc is toggled. calling updatepowerstate() when rbc changes state fixes this.

2) Originally, when rbc was toggled, due to the colour filter and therefore change to nits mapping,
the backlight value would compensate for the changes, causing the backlight & slider to increase
when extra dim was turned on, and vice versa. This is very unintuitive to the user.
This change sets the old brightness value when rbc is toggled ensuring the screen does darken
as the user would expect it to when they turn on this feature.

Bug: 189819485
Test: atest FrameworksServicesTests:ReduceBrightColorsTintControllerTest
Change-Id: I4ae0baca44757360e049ab71415b5a518f4999ba