Remove controls from carousel when not in manager
If the user is in a bad state where controls exist in the carousel but
their keys have been removed from the manager, allow the dismiss button
to remove the player from the carousel directly.

The device isn't expected to get into this state normally but this will
allow users to remove the bad controls without needing to reboot.
Also added dump to MediaCarouselController to help debug future reports.

Bug: 190799184
Test: atest MediaDataManagerTest MediaControlPanelTest
Test: take bugreport, verify MediaCarouselController data is in dumpsys
Test: manual - force bad controls state, verify can still remove
Change-Id: I7af130c90c7b661cc7cc8bbeb3ca1fd5d81c8913