Allow task fragment to use higher animation level
The method canCreateRemoteAnimationTarget was used to decide
whether to promote the animation level for transition. E.g.
If 2 activities in task fragment of the same task are opening,
the animation target should be the task instead of the 2
activities individually.

The method createRemoteAnimationTarget is also moved from Task
to TaskFragment for consistency of the ability to create the
animation target.

Test: Move a task with embedded activities to front.
Bug: 194673563
Change-Id: I51aed380d0a20ce7981bf5d1911dfa82d8b54165