Center bubbles horizontally on large screens
Previously bubbles were anchored to the edge they were
expanded on, with the expanded view next to it.

This CL modifies it such that the bubble expanded view
will be centered, and the bubbles will be anchored to
the expanded view.

BubblePositioner is updated to return the x,y of the
bubble. This allows an x value to be specified
to inset the bubble positions on large screens.

Updates phone landscape layout to have padding at the
top (to match with tablet layouts). The bubble
overflow is modified to be full height with top &
bottom padding in phone landscape.

Fixes an issue where we were scaling more than needed
in landscape.

Test: manual - check expanded layout in phone, phone
               landscape, and tablet.
             - check with gesture nav & 3 button
             - check everything looks fine after a
               display size & orientation change
Bug: 193910391
Change-Id: Iba4a3ac0932f716eea929fc0d8e8ef7f1b1e3f5e