Fix DisplayDeviceConfig crash for values > 255
If screenBrightnessBacklight values in a config.xml file go above 255,
they leave a string of 1.0 values for every one that is above that. This
change ensures that this does not happen, and clamps the array so that
it contains a maximum of 1 value of 1.0.

Bug: 197491062
Test: manual - change config.xml config_screenBrightnessBacklight to go above 255
Change-Id: Id245ea902f265aa86a2caeff9d24a67fad4c37f7
(cherry picked from commit 357286666194cc35002d1332cb603a3d1d18dfc2)