Hide security footer container (QSBH) in portrait
In order to have both sides of the notch protection be of the same size
(and the notch protection be centered), we need to hide the other
container in the date_privacy header. This is fine to hide in portrait
as it should only have content in landscape.

Test: manual, using display cutout and checking layout dump
Fixes: 197768177
Change-Id: I4e5754c92c386cb9ce9e2ecc7e22a0f0c0c3fd04
Merged-In: I4e5754c92c386cb9ce9e2ecc7e22a0f0c0c3fd04