Check the transient state by using public types
Previously, we only checked if ITYPE_STATUS_BAR or ITYPE_NAVIGATION_BAR
is showing transiently or not when deciding the control target. There
could be an issue if the system doesn't have ITYPE_NAVIGATION_BAR but
has ITYPE_EXTRA_NAVIGATION_BAR. When that happens, mDummyControlTarget
in InsetsPolicy cannot be the navigation control target, so it cannot
show the transient bar.

This CL uses public types to decide the control target in transient
mode.

Bug: 198193735
Test: 1. Enable task bar from DeviceProfile.
      2. Set smallest width to 600 dp from developer options.
      3. Open an app which hides all system bars in sticky-immersive
         mode.
      4. Swipe from top and see if task bar is visible.
Change-Id: Id0a0b60831e11e32fbeac93330144535ed7cb416