Include permission-based notification allow/bans in dumps.
Notification permission info is read from the permission helper and combined across all users on the device, which is then passed to the various dump/stats collection methods.

This change covers dumpImpl (for logcat), dumpJson, and dumpProto. It also modifies the notification pulled atoms (PackageNotificationPreferences) to read from the app permissions if needed.

Bug: 194833441
Test: atest NotificationManagerServiceTest, NotificationPermissionMigrationTest, PreferencesHelperTest
Change-Id: I3b5d8e4a31f1a183a5404b2a6c256ea54e0da270