Visible activity can be an ime target
While an activity is still visible, it is still an ime target
even if it is going away.

This fixes the problem with shell transitions where IME is
detached from the live tile while entering recents. During recents,
the the recents activity is on-top but has canBeImeTarget() -> false.
The IME code then searches for the next-highest ImeTarget candidate.
Since the live-tile activity is still visible, it should be the
next candidate.

Bug: 193565597
Test: open app and open IME. Go to recents. IME should be attached
      to app
      atest WindowStateTests
Change-Id: I197359e2c21da4c4f1dedaaa43fd1ca76a135ce8