ShellTranstion: Apply recents input consumer when the window in transion
Like legacy recents animation will check if the focused window is
animating with recents animation, if so, requests the input focus on
recents input consumer, if not, means the window is not controlled by
recents animation and we should give it focus.
(e.g. Tapping on StatusBar)

Otherwise the input focus will be consumed by recents input consumer
and then tapping the editor on direct-reply notification will not able
to show IME.

Bug: 210390667
Test: manual as steps
1) adb shell setprop persist.debug.shell_transit 1)
2) install EditTextVariations.apk
3) launch EditTextVariations, choose "Direct Reply" to popup
   directly-reply notification
4) launch Android Message with focusing the editor to show IME
5) swipe up to enter the overview screen
6) swipe down the notification shade with pressing "Direct Reply Text"
   on "Message from UserHandle()" notification.
7) Expect IME should show up

Change-Id: I8830f2deaa623985b05a5fc85d2a196c5cfbf81d