Avoid calling the same on change listeners twice
Currently the association store implementation notifies every
OnChangeListener twice: once with a generic onAssociationChanged call
and then with a more specialized call depending on the nature of the
change.

This CL moves the responsibility to decide what kind of change this
should be into the default implementation of the generic
onAssociationChanged method in the interface.
This still allows to override this behavior while alleviating the
need to handle double callbacks in the specialised callback in all
listener implementations.

Test: atest CtsCompanionDeviceManagerCoreTestCases
Test: atest CtsCompanionDeviceManagerUiAutomationTestCases
Change-Id: Idd1150c096ff2e55d805edd2c4c34a2a86ecf7fe