Temporarily disable some biometric prompt tests.
The removal cl (https://googleplex-android-review.git.corp.google.com/c/platform/frameworks/base/+/16980367/3) introduced an issue with the tests that causes an indirect failure for the systemui pre-submit tests, which does not repo locally. This should be reverted as soon as possible and is only temporarily ignored to enable dogfood builds.

Bug: 222119218
Test: presubmit

Change-Id: I41d64ffc528bccc98a2e5da7a98edf472539f3de