Fixs the ANR issue
When doing the drag and drop an app from the Taskbar, the dragState
will hold the wm global lock to sync the input state through the
surface flinger which callbacks the windows through the
onWindowInfoChanged. If we also hold the same lock in the
onWindowInfoChanged, then the ANR will happen.

To fix this ANR issue, we need to post a message and then handle
the windowInfo data.

Bug: 221497060
Test: a11y CTS & unit tests
Change-Id: Ib655546f5810e357fa45419ece32e9e742b6506c