Restrict device wakeup via PowerManager.ACQUIRE_CAUSES_WAKEUP.
The default value of OP_TURN_SCREEN_ON is still MODE_ALLOWED, which
means the user needs to actively disallow the special app access for an
app to lose the functionality of turning the screen on via wakelock.
Only wake the power group linked to the wake lock if valid power group
id is provided.

Bug: 216114297
Bug: 219568783
Test: manual
Change-Id: Ib6df6e6c4f71adbb9bc8fd3db80708e4af0c40cc