Add support for font-related classes to RNR
Also, add some special case logic to RegisterMethodsOrDie to handle
registration for classes in the package 'org/robolectric/nativeruntime'.
This was done because there are two 'FontFamily' classes with native
methods: android.graphics.FontFamily, and
android.graphics.fonts.FontFamily. To support this case, the JNI
registration was android.graphics.fonts.FontFamily was explicitly
updated to register to the 'FontsFontFileNatives' java class.

Test: m -j robolectric_native_runtime dist
Change-Id: I6071aa4d20f623ac76f24aa369c5d0a30a21926c