Compile against the static variant of libnativehelper
libnativehelper_lazy was designed to reduce binary size by dynamically
loading libnativehelper when it is requested. Update the
robolectric_native_runtime target to use the static variant of
libnativehelper.

Also, enable BitmapFactory.nativeDecodeFileDescriptor, as it works fine
on Linux. This is what requires libnativehelper. For Windows we can use
a workaround, such as reading all the contents of the file into a
ByteBuffer.

Test: m -j robolectric_native_runtime dist

Change-Id: Ie22520737d82fc48da7949f4d112c63c57eb8232