Fix composed VibrationEffect timing out
Fix timeout of playback of composed vibration effects (not only
primitives). This happens because the timings used for triggering the
next step after a HAL callback is received is propagaing the extra
timeout given to the current step to wait before playing it, causing the
consecutive steps to wait for a timeout no longer necessary.

This is being tested now in VibrationThreadTest, with a composition of
effects and primitives that has a realtime replacement with fallback,
and in VibrationManagerServiceTest, that is using blocking vibrate calls
to avoid some flakiness also found in those tests.

Fix: 238311929
Test: VibrationManagerServiceTest & VibrationThreadTest
Change-Id: I15ec5b977076903660b676ef671a472c10d9c30b