Apply navigation bars insets to the fullscreen dialog
This CL ensures that we apply the navigation/task bars instead of
fullscreen dialogs animated by DialogLaunchAnimator to the fullscreen
View (that is the parent of the dialog content + background) and not to
the Dialog DecorView. This prevents the dialog from being clipped by the
navigation/task bars insets when animating.

Bug: 232063468
Test: Manual
Change-Id: I3e98205a2b3da0f5536a48601b39620e7933b9c9