Reduce blocking calls to Settings in main thread
In most cases, mTileList contains the same information as the Settings
value, so there's no need to retrieve it before modifying it. Keep track
of that with a dirty flag.

This way, we reduce the number of blocking calls in the main thread (as
that's the thread that processes the tiles).

Test: manual, factory reset
Test: existing QSTileHostTest
Test: performance metrics are back to baseline
Fixes: 240256263
Change-Id: Idebd37d1458c80330b60802729575219b6a7b49a
Merged-In: Idebd37d1458c80330b60802729575219b6a7b49a
(cherry picked from commit 4da071e035b647f498936bdb2b557aa08bd9df0f)
Merged-In: Idebd37d1458c80330b60802729575219b6a7b49a